penrose
Module builtin

Built-in functionality for penrose.

The code in this module is built using the APIs available to users and provided as a minimal “out of the box” experience for using the window manager.

More advanced functionality is available via the extensions module.

Modules

actions

Helpers and pre-defined actions for use in user defined key bindings

hooks

Built-in hooks

layout

Built-in layouts.