penrose::builtin
Module actions

Helpers and pre-defined actions for use in user defined key bindings

Modules

floating

Actions for manipulating floating windows.

Functions

broadcast_layout_message

Send a message to all layouts available to the current workspace

exit

Exit penrose

key_handler

Construct a KeyEventHandler from a closure or free function

log_current_state

Info log the current window manager State for debugging purposes.

modify_with

Mutate the ClientSet and refresh the on screen state

remove_and_unmap_focused_client

Remove the currently focused client from state and unmap it WITHOUT closing the client program. This is provided for removing clients that have been accidentally tiled when they should have been ignored.

send_layout_message

Send a message to the currently active layout

spawn

Spawn an external program as part of a key binding