penrose::builtin::layout

Module messages

Built-in layout messages.

It is not a hard requirement for Layout implementations to handle each of the messages provided by this module but wherever possible you should attempt to do so if the semantics of the message make sense for the layout you are writing.

Structs

ExpandMain

Expand the size of the main area of the Layout.

Hide

A Message sent when a Layout is no longer visible (e.g. Layout changed on a visible Workspace or the workspace itself becoming hidden).

IncMain

Alter the number of clients contained in the main area of the Layout.

Mirror

Mirror the Layout over either the horizontal or vertical axis.

Rotate

Rotate the Layout to a new orientation

ShrinkMain

Shrink the size of the main area of the Layout

UnwrapTransformer

Unwrap a LayoutTransformer to return the underlying Layout.