Struct penrose::builtin::layout::transformers::Gaps

pub struct Gaps {
    pub layout: Box<dyn Layout>,
    pub outer_px: u32,
    pub inner_px: u32,
}

Simple gaps around the window placement of the enclosed Layout.

outer_px controls the width of the gap around the edge of the screen and inner_px controls the gap around each individual window. Set both equal to one another to have a consistant gap size in all places.

Fields
layout: Box<dyn Layout>

The inner Layout having gaps applied to it.

outer_px: u32

The desired outer gap size in pixels

inner_px: u32

The desired inner gap size in pixels

Implementations

impl Gaps

pub fn wrap( layout: Box<dyn Layout>, outer_px: u32, inner_px: u32 ) -> Box<dyn Layout>

Wrap an existing Layout with the given gap sizes.