Module penrose::contrib::extensions

Functionality extensions for penrose

Most of these extension work by spawning and / or managing external programs as a sub-process.

Modules

dmenu
A simple wrapper for suckless' dmenu tool for providing quick text based menus

notify_send
A simple wrapper around notify-send to allow for generating notifications

scratchpad
A scratchpad that holds a single client

Structs

DMenu
A wrapper around the suckless dmenu program for creating dynamic menus in penrose.

DMenuConfig
Config for running a DMenu selection

NotifyConfig
Notification configuration

Scratchpad
Spawn and manage a single Client which can then be shown above the current layout.

Enums

MenuMatch
The result of attempting to match against user input in a DMenu

NotifyLevel
A notification level when calling notify-send

Functions

notify_send
Send a notification using the notify-send command line program

notify_send_error_handler
A simple error handler that uses 'notify-send' to display a dialog window with the error message.