Module penrose::contrib::hooks

Additional common hooks that can be used out of the box with minimal config.

Structs

ActiveClientAsRootName
Automatically set the X root window WM_NAME property to be the WM_NAME of the active window.

AutoSetMonitorsViaXrandr
Automatically set the current monitors and their positions whenever there is an xrandr change

ClientSpawnRules
Move clients with a matching WM_NAME to a target workspace when they are spawned.

DefaultWorkspace
Whenever a focus moves to the workspace 'name' and the workspace is empty, set a specific layout and spawn a set of default clients.

LayoutSymbolAsRootName
Automatically set the X root window WM_NAME property to be the current layout symbol for the active workspace.

RemoveEmptyWorkspaces
Automatically remove empty workspaces when they lose focus.

Enums

SpawnRule
An individual workspace mapping for ClientSpawnRules