penrose::core
Module bindings

Setting up and responding to user defined key/mouse bindings

Structs

KeyCode

A key press and held modifiers

ModifierKeyIter

An iterator over the variants of ModifierKey

MotionNotifyEvent

Mouse motion with a held button and optional modifiers

MouseEvent

A mouse movement or button event

MouseEventData

Data from a button press or motion-notify event

MouseState

A mouse state specification indicating the button and modifiers held

Enums

KeyPress

Abstraction layer for working with key presses

ModifierKey

Known modifier keys for bindings

MouseButton

Known mouse buttons for binding actions

MouseEventKind

The types of mouse events represented by a MouseEvent

Traits

KeyEventHandler

Some action to be run by a user key binding

MouseEventHandler

An action to be run in response to a mouse event

Functions

click_handler

Convert a KeyEventHandler to a MouseEventHandler that runs on button Press events.

keycodes_from_xmodmap

Run the xmodmap command to dump the system keymap table.

parse_keybindings_with_xmodmap

Parse string format key bindings into KeyCode based KeyBindings using the command line xmodmap utility.

Type Aliases

KeyBindings

User defined key bindings

KeyCodeMask

A u16 X key-code bitmask

KeyCodeValue

A u8 X key-code enum value

MouseBindings

User defined mouse bindings