penrose::core::bindings

Function click_handler

pub fn click_handler<X: XConn + 'static>(
    kh: Box<dyn KeyEventHandler<X>>,
) -> Box<dyn MouseEventHandler<X>>

Convert a KeyEventHandler to a MouseEventHandler that runs on button Press events.

This allows for running pre-existing simple key handlers that do not care about press/release or motion behaviour as a simplified mouse event handler.

Example

use penrose::builtin::actions::floating::sink_all;
use penrose::core::bindings::{click_handler, MouseEventHandler};
use penrose::x11rb::RustConn;

let handler: Box<dyn MouseEventHandler<RustConn>> =  click_handler(sink_all());