penrose::core::bindings

Function keycodes_from_xmodmap

pub fn keycodes_from_xmodmap() -> Result<HashMap<String, u8>>

Run the xmodmap command to dump the system keymap table.

This is done in a form that we can load in and convert back to key codes. This lets the user define key bindings in the way that they would expect while also ensuring that it is easy to debug any odd issues with bindings by referring the user to the xmodmap output.

Panics

This function will panic if it is unable to fetch keycodes using the xmodmap binary on your system or if the output of xmodmap -pke is not valid