penrose::core::bindings

Trait KeyEventHandler

pub trait KeyEventHandler<X>
where
    X: XConn,
{
    // Required method
    fn call(&mut self, state: &mut State<X>, x: &X) -> Result<()>;
}

Some action to be run by a user key binding

Required Methods

fn call(&mut self, state: &mut State<X>, x: &X) -> Result<()>

Call this handler with the current window manager state