penrose::core::bindings

Trait MouseEventHandler

pub trait MouseEventHandler<X>
where
    X: XConn,
{
    // Required methods
    fn on_mouse_event(
        &mut self,
        evt: &MouseEvent,
        state: &mut State<X>,
        x: &X,
    ) -> Result<()>;
    fn on_motion(
        &mut self,
        evt: &MotionNotifyEvent,
        state: &mut State<X>,
        x: &X,
    ) -> Result<()>;
}

An action to be run in response to a mouse event

Required Methods

fn on_mouse_event( &mut self, evt: &MouseEvent, state: &mut State<X>, x: &X, ) -> Result<()>

Called when the MouseState associated with this handler is seen with a button press or release.

fn on_motion( &mut self, evt: &MotionNotifyEvent, state: &mut State<X>, x: &X, ) -> Result<()>

Called when the ModifierKeys associated with this handler are seen when the mouse is moving.