penrose::core::conn

Function manage_without_refresh

pub fn manage_without_refresh<C: Conn>(
    id: WinId,
    tag: Option<&str>,
    state: &mut State<C>,
    conn: &mut C,
) -> Result<()>

The main logic for inserting a new client into the StackSet without any refresh of the X state. In normal window manager operation, the manage method on XConnExt is always used: this is provided independently to support managing existing clients on startup.