penrose::core::conn

Trait Conn

pub trait Conn: Send + Sized {
    type Event: ConnEvent;
    type State: Debug + Send + Sized;
    type KeyBindingKey: Debug + Copy + Clone + PartialEq + Eq + Hash + Send + Sized;

    // Required methods
    fn initial_state(&mut self) -> Self::State;
    fn root(&mut self) -> WinId;
    fn next_event(&mut self) -> Result<Self::Event>;
    fn handle_event(
        &mut self,
        evt: Self::Event,
        key_bindings: &mut KeyBindings<Self>,
        mouse_bindings: &mut MouseBindings<Self>,
        state: &mut State<Self>,
    ) -> Result<()>;
    fn flush(&mut self);
    fn grab(
        &mut self,
        key_codes: &[Self::KeyBindingKey],
        mouse_states: &[MouseState],
    ) -> Result<()>;
    fn existing_clients(&mut self) -> Result<Vec<WinId>>;
    fn manage_existing_clients(&mut self, state: &mut State<Self>) -> Result<()>;
    fn screen_details(&mut self) -> Result<Vec<Rect>>;
    fn cursor_position(&mut self) -> Result<Point>;
    fn warp_pointer(&mut self, id: WinId, x: i16, y: i16) -> Result<()>;
    fn position_client(&mut self, id: WinId, r: Rect) -> Result<()>;
    fn show_client(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>;
    fn hide_client(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>;
    fn withdraw_client(&mut self, id: WinId) -> Result<()>;
    fn kill_client(&mut self, id: WinId) -> Result<()>;
    fn focus_client(&mut self, id: WinId) -> Result<()>;
    fn client_geometry(&mut self, id: WinId) -> Result<Rect>;
    fn client_title(&mut self, id: WinId) -> Result<String>;
    fn client_pid(&mut self, id: WinId) -> Option<u32>;
    fn client_should_float(
        &mut self,
        id: WinId,
        floating_classes: &[String],
    ) -> bool;
    fn client_should_be_managed(&mut self, id: WinId) -> bool;
    fn client_is_fullscreen(&mut self, id: WinId) -> bool;
    fn client_transient_parent(&mut self, id: WinId) -> Option<WinId>;
    fn set_client_border_color(
        &mut self,
        id: WinId,
        color: impl Into<Color>,
    ) -> Result<()>;
    fn set_initial_properties(
        &mut self,
        id: WinId,
        config: &Config<Self>,
    ) -> Result<()>;
    fn restack<'a, I>(&mut self, ids: I) -> Result<()>
       where WinId: 'a,
             I: Iterator<Item = &'a WinId>;

}

A platform agnostic backing connection

Required Associated Types

type Event: ConnEvent

The event type used by this connection

type State: Debug + Send + Sized

Additional data that will be kept separately within the main State.

This is typically used to be able to support having generic shared state when writing families of Conn implementations such as XConn.

type KeyBindingKey: Debug + Copy + Clone + PartialEq + Eq + Hash + Send + Sized

The type that is used as the key for a KeyBindings map using this Conn.

Required Methods

fn initial_state(&mut self) -> Self::State

Called once when the window manager state is first created.

fn root(&mut self) -> WinId

The ID of the window manager root window.

fn next_event(&mut self) -> Result<Self::Event>

Block and wait for the next event so it can be processed.

fn handle_event( &mut self, evt: Self::Event, key_bindings: &mut KeyBindings<Self>, mouse_bindings: &mut MouseBindings<Self>, state: &mut State<Self>, ) -> Result<()>

Process the an event

fn flush(&mut self)

Flush any pending events to the underlying back end.

fn grab( &mut self, key_codes: &[Self::KeyBindingKey], mouse_states: &[MouseState], ) -> Result<()>

Grab the specified key and mouse states, intercepting them for processing within the window manager itself.

fn existing_clients(&mut self) -> Result<Vec<WinId>>

Ask the X server for the IDs of all currently known client windows

fn manage_existing_clients(&mut self, state: &mut State<Self>) -> Result<()>

Request a client windows’s current workspace

fn screen_details(&mut self) -> Result<Vec<Rect>>

The dimensions of each currently available screen.

fn cursor_position(&mut self) -> Result<Point>

The current (x, y) coordinate of the mouse cursor.

fn warp_pointer(&mut self, id: WinId, x: i16, y: i16) -> Result<()>

Reposition the mouse cursor to the given (x, y) coordinates within the specified window.

fn position_client(&mut self, id: WinId, r: Rect) -> Result<()>

Update the geometry of a given client based on the given Rect.

fn show_client(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>

Display a client on the screen at its current position.

fn hide_client(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>

Hide a client

fn withdraw_client(&mut self, id: WinId) -> Result<()>

Withdraw a client

fn kill_client(&mut self, id: WinId) -> Result<()>

Kill the given client window, closing it.

fn focus_client(&mut self, id: WinId) -> Result<()>

Set input focus to be held by the given client window.

fn client_geometry(&mut self, id: WinId) -> Result<Rect>

Look up the current dimensions and position of a given client window.

fn client_title(&mut self, id: WinId) -> Result<String>

Request the title of a given client window.

fn client_pid(&mut self, id: WinId) -> Option<u32>

Request a window’s PID.

fn client_should_float( &mut self, id: WinId, floating_classes: &[String], ) -> bool

Check whether or not the given client should be assigned floating status or not.

fn client_should_be_managed(&mut self, id: WinId) -> bool

For a given existing client being processed on startup, determine whether we need to bring it into our internal state and manage it.

fn client_is_fullscreen(&mut self, id: WinId) -> bool

Check whether this client is currently fullscreen or not

fn client_transient_parent(&mut self, id: WinId) -> Option<WinId>

The id of the parent for this client if it is transient

fn set_client_border_color( &mut self, id: WinId, color: impl Into<Color>, ) -> Result<()>

Update the border color of the given client window.

fn set_initial_properties( &mut self, id: WinId, config: &Config<Self>, ) -> Result<()>

Set the initial window properties for a newly managed window.

fn restack<'a, I>(&mut self, ids: I) -> Result<()>
where WinId: 'a, I: Iterator<Item = &'a WinId>,

Restack the given windows, each one above the last.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.