penrose::core::conn

Trait ConnExt

pub trait ConnExt: Conn + Sized {
    // Provided methods
    fn kill_focused(&mut self, state: &mut State<Self>) -> Result<()> { ... }
    fn manage(&mut self, id: WinId, state: &mut State<Self>) -> Result<()> { ... }
    fn unmanage(&mut self, id: WinId, state: &mut State<Self>) -> Result<()> { ... }
    fn modify_and_refresh<F>(
        &mut self,
        state: &mut State<Self>,
        f: F,
    ) -> Result<()>
       where F: FnMut(&mut StackSet<WinId>) { ... }
    fn refresh(&mut self, state: &mut State<Self>) -> Result<()> { ... }
    fn position_clients(&mut self, state: &State<Self>) -> Result<()> { ... }
    fn set_active_client(
        &mut self,
        id: WinId,
        state: &mut State<Self>,
    ) -> Result<()> { ... }
    fn warp_pointer_to_window(&mut self, id: WinId) -> Result<()> { ... }
    fn warp_pointer_to_screen(
        &mut self,
        state: &mut State<Self>,
        screen_index: usize,
    ) -> Result<()> { ... }
    fn query(&mut self, query: &dyn Query<Self>, id: WinId) -> Result<bool> { ... }
    fn query_or(
        &mut self,
        default: bool,
        query: &dyn Query<Self>,
        id: WinId,
    ) -> bool { ... }
}

Extended functionality for Conn impls in order to run the window manager.

Provided Methods

fn kill_focused(&mut self, state: &mut State<Self>) -> Result<()>

Kill the focused client if there is one

fn manage(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>

Establish the window manager state for the given client window and refresh the current X state.

fn unmanage(&mut self, id: WinId, state: &mut State<Self>) -> Result<()>

Remove the window manager state for the given client window and refresh the current X state.

fn modify_and_refresh<F>(&mut self, state: &mut State<Self>, f: F) -> Result<()>
where F: FnMut(&mut StackSet<WinId>),

Apply a pure function that modifies a [ClientSet] and then handle refreshing the WindowManager state and associated X11 calls.

This is the main logic that drives what the user will see on the screen in terms of window placement, focus and borders. Everything is driven from a diff of the pure ClientSet state before and after some mutating operation that was carried out by f.

fn refresh(&mut self, state: &mut State<Self>) -> Result<()>

Refresh the current X server state based on a diff of the current state against the state when we last refreshed.

fn position_clients(&mut self, state: &State<Self>) -> Result<()>

Restack and set the geometry for an ordered list of client windows and their associated positions. The provided positions are shrunk by the current border size in order to position the windows correctly within the frame given by the border.

See restack for details of stacking order is determined.

fn set_active_client( &mut self, id: WinId, state: &mut State<Self>, ) -> Result<()>

Update the currently focused client and refresh the X state.

fn warp_pointer_to_window(&mut self, id: WinId) -> Result<()>

Warp the mouse cursor to the center of the given client window.

fn warp_pointer_to_screen( &mut self, state: &mut State<Self>, screen_index: usize, ) -> Result<()>

Warp the mouse cursor to the center of the given screen.

fn query(&mut self, query: &dyn Query<Self>, id: WinId) -> Result<bool>

Run the provided Query, returning the result.

fn query_or( &mut self, default: bool, query: &dyn Query<Self>, id: WinId, ) -> bool

Run the provided Query, returning the result or a default value if there were any errors encountered when communicating with the X server.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.