Module penrose::core::data_types

Simple data types and enums

Structs

Point
An x,y coordinate pair

Region
An X window / screen position: top left corner + extent

Enums

Border
X window border kind

Change
Increment / decrement a value

PropVal
A client propert value that can be set.

RelativePosition
A relative position along the horizontal and vertical axes

WinAttr
Window attributes for an X11 client window (not all are curently implemented)

WinConfig
Config options for X windows (not all are currently implemented)

WinType
A window type to be specified when creating a new window in the X server

Type Definitions

ResizeAction
Output of a Layout function: the new position a window should take

WinId
An X window ID