Module penrose::core::helpers

Utility functions for use in other parts of penrose

Functions

index_selectors
Create a Vec of index selectors for the given input slice

keycodes_from_xmodmap
Run the xmodmap command to dump the system keymap table.

logging_error_handler
A simple error handler that just logs the error to the penrose log stream

spawn
Run an external command

spawn_for_output
Run an external command and return its output.

spawn_with_args
Run an external command with the specified command line arguments