Trait penrose::core::hooks::LayoutHook

pub trait LayoutHook<X>
where
    X: XConn,
{
    // Provided methods
    fn transform_initial_for_screen(
        &mut self,
        screen_index: usize,
        r: Rect,
        state: &State<X>,
        x: &X,
    ) -> Rect { ... }
    fn transform_initial(&mut self, r: Rect, state: &State<X>, x: &X) -> Rect { ... }
    fn transform_positions_for_screen(
        &mut self,
        screen_index: usize,
        r: Rect,
        positions: Vec<(Xid, Rect)>,
        state: &State<X>,
        x: &X,
    ) -> Vec<(Xid, Rect)> { ... }
    fn transform_positions(
        &mut self,
        r: Rect,
        positions: Vec<(Xid, Rect)>,
        state: &State<X>,
        x: &X,
    ) -> Vec<(Xid, Rect)> { ... }
    fn then<H>(self, next: H) -> ComposedLayoutHook<X>
       where H: LayoutHook<X> + 'static,
             Self: Sized + 'static { ... }
    fn boxed(self) -> Box<dyn LayoutHook<X>>
       where Self: Sized + 'static { ... }
    fn then_boxed(self, next: Box<dyn LayoutHook<X>>) -> Box<dyn LayoutHook<X>>
       where Self: Sized + 'static,
             X: 'static { ... }
}

Logic to run before and after laying out clients

Provided Methods

fn transform_initial_for_screen( &mut self, screen_index: usize, r: Rect, state: &State<X>, x: &X, ) -> Rect

Optionally modify the screen dimensions being given to a Layout on a particular screen index.

By default this just calls through to LayoutHook::transform_initial.

fn transform_initial(&mut self, r: Rect, state: &State<X>, x: &X) -> Rect

Optionally modify the screen dimensions being given to a Layout

fn transform_positions_for_screen( &mut self, screen_index: usize, r: Rect, positions: Vec<(Xid, Rect)>, state: &State<X>, x: &X, ) -> Vec<(Xid, Rect)>

Optionally modify the client positions returned by a Layout on a particular screen index.

By default this just calls through to LayoutHook::transform_positions.

fn transform_positions( &mut self, r: Rect, positions: Vec<(Xid, Rect)>, state: &State<X>, x: &X, ) -> Vec<(Xid, Rect)>

Optionally modify the client positions returned by a Layout

fn then<H>(self, next: H) -> ComposedLayoutHook<X>
where H: LayoutHook<X> + 'static, Self: Sized + 'static,

Compose this hook with another LayoutHook.

fn boxed(self) -> Box<dyn LayoutHook<X>>
where Self: Sized + 'static,

Convert to a trait object

fn then_boxed(self, next: Box<dyn LayoutHook<X>>) -> Box<dyn LayoutHook<X>>
where Self: Sized + 'static, X: 'static,

Compose this hook with a boxed LayoutHook.