Function penrose::core::layout::monocle

pub fn monocle(
    clients: &[&Client], 
    focused: Option<WinId>, 
    monitor_region: &Region, 
    _: u32, 
    _: f32
) -> Vec<ResizeAction>

A simple monolve layout that places uses the maximum available space for the focused client and unmaps all other windows.