Trait penrose::core::layout::Layout

pub trait Layout {
    // Required methods
    fn name(&self) -> String;
    fn boxed_clone(&self) -> Box<dyn Layout>;
    fn layout(
        &mut self,
        s: &Stack<Xid>,
        r: Rect
    ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>);
    fn handle_message(&mut self, m: &Message) -> Option<Box<dyn Layout>>;

    // Provided methods
    fn boxed(self) -> Box<dyn Layout>
       where Self: Sized + 'static { ... }
    fn layout_workspace(
        &mut self,
        tag: &str,
        stack: &Option<Stack<Xid>>,
        r: Rect
    ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>) { ... }
    fn layout_empty(
        &mut self,
        r: Rect
    ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>) { ... }
}

A Layout is responsible for positioning a Stack of clients in a given coordinate space denoting the dimensions of users display.

Mutating the state of a Layout is possible by sending it a Message which can then either modify the existing layout (e.g. increase the number of clients positioned in a “main” area) or replace the existing Layout with a new one. There is no requirement to be able to handle all message types.

Required Methods

fn name(&self) -> String

A short display name for this Layout, appropriate for rendering in a status bar as an indicator of which layout is currently being used.

fn boxed_clone(&self) -> Box<dyn Layout>

Provide a clone of this Layout wrapped as a trait object. (Trait objects can not require Clone directly)

fn layout( &mut self, s: &Stack<Xid>, r: Rect ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>)

Generate screen positions for clients from a given Stack.

See Layout::layout_workspace for details of how positions should be returned.

fn handle_message(&mut self, m: &Message) -> Option<Box<dyn Layout>>

Process a dynamic Message.

See the trait level docs for details on what is possible with messages.

Provided Methods

fn boxed(self) -> Box<dyn Layout>where Self: Sized + 'static,

Convert to a trait object.

fn layout_workspace( &mut self, tag: &str, stack: &Option<Stack<Xid>>, r: Rect ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>)

Generate screen positions for clients on a given crate::pure::Workspace.

If you do not need to know the details of the workspace being laid out, you should can use the default implementation of this methods which calls Layout::layout if there are any clients present and Layout::layout_empty if not.

Positioning clients

For each client that should be shown on the screen a pair of its Xid and a Rect should be provided, indicating the screen position the client should be placed at. To hide a client that was present in the crate::pure::Workspace simply do not provide a position for it. (You may also provide positions for clients that were not present in the input if you have the Xid available.)

The order in which the (Xid, Rect) pairs are returned determines the stacking order on the screen. It does not have to match the stack order of the clients within the crate::pure::Workspace.

Returning a new layout

When a layout is run it may optionally replace itself with a new Layout. If Some(layout) is returned from this method, it will be swapped out for the current one after the provided positions have been applied.

fn layout_empty( &mut self, r: Rect ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>)

Generate screen positions for an empty Stack.

See Layout::layout_workspace for details of how positions should be returned.