Struct penrose::core::manager::WindowManager

pub struct WindowManager<X: XConn> { /* fields omitted */ }

WindowManager is the primary struct / owner of the event loop for penrose.

It handles most (if not all) of the communication with the underlying XConn, responding to XEvents emitted by it. User key / mouse bindings are parsed and bound on the call to grab_keys_and_run and then triggered when corresponding XEvent instances come through in the main event loop.

A note on examples

The following examples are for intended for reference purposes only and are not all run as doc-tests due to the requirement for a connection to the X server. For further details and example use, please see the examples directory in the penrose git repo.

Implementations

impl<X: XConn> WindowManager<X>

pub fn new(
config: Config,
conn: X,
hooks: Hooks<X>,
error_handler: ErrorHandler
) -> Self

Construct a new window manager instance using a chosen XConn backed to communicate with the X server.

Example

use penrose::{
    core::{Config, WindowManager},
    xcb::XcbConnection,
    logging_error_handler
};

let mut wm = WindowManager::new(
    Config::default(),
    XcbConnection::new().unwrap(),
    vec![],
    logging_error_handler(),
);

// additional set-up logic here

if let Err(e) = wm.init() {
    panic!("failed to initialise WindowManager: {}", e);
}
wm.log("ready to call grab_keys_and_run!").unwrap();

pub fn hydrate_and_init(
&mut self,
hooks: Hooks<X>,
error_handler: ErrorHandler,
layout_funcs: HashMap<&str, LayoutFunc>
) -> Result<()>

Restore missing state following serde deserialization.

Errors

The deserialized state will be checked and validated for internal consistency and consistency with the current X server state using the deserialized XConn. If the state is not a valid snapshot then an error will be returned. Examples of invalid state include:

Not providing a required layout function in layout_funcs

Workspace Client IDs not appearing in the WindowManager client_map

Being unable to connect to the X Server

Example

use penrose::{
    contrib::hooks::{SpawnRule, ClientSpawnRules},
    core::{
        hooks::Hooks,
        layout::{floating, side_stack, LayoutFunc},
        manager::WindowManager,
    },
    xcb::XcbConnection,
    logging_error_handler
};

// Hooks that we want to set up on restart
let hooks: Hooks<_> = vec![
    ClientSpawnRules::new(vec![
        SpawnRule::ClassName("xterm-256color" , 3),
        SpawnRule::WMName("Firefox Developer Edition" , 7),
    ])
];

// The layout functions we were using previously
let layout_funcs = map! {
    "[side]" => side_stack as LayoutFunc,
    "[----]" => floating as LayoutFunc,
};

let json_str = "...";  // Load in the serialized state from somewhere
let mut manager: WindowManager<XcbConnection> = serde_json::from_str(&json_str).unwrap();
assert!(manager.hydrate_and_init(hooks, logging_error_handler(), layout_funcs).is_ok());

pub fn init(&mut self) -> Result<()>

This initialises the WindowManager internal state but does not start processing any events from the X server. If you need to perform any custom setup logic with the WindowManager itself, it should be run after calling this method and before WindowManager::grab_keys_and_run.

Example

use penrose::{
    core::{Config, WindowManager},
    xcb::XcbConnection,
    logging_error_handler
};

let mut wm = WindowManager::new(
    Config::default(),
    XcbConnection::new().unwrap(),
    vec![],
    logging_error_handler(),
);

// additional set-up logic here

if let Err(e) = wm.init() {
    panic!("failed to initialise WindowManager: {}", e);
}
wm.log("ready to call grab_keys_and_run!").unwrap();

pub fn grab_keys_and_run(
&mut self,
key_bindings: KeyBindings<X>,
mouse_bindings: MouseBindings<X>
) -> Result<()>

This is the main event loop for the WindowManager.

The XConn::wait_for_event method is called to fetch the next event from the X server, after which it is processed into a set of internal EventActions which are then processed by the WindowManager to update state and perform actions. This method is an infinite loop until the WindowManager::exit method is called, which triggers XConn::cleanup before exiting the loop. You can provide any additional teardown logic you need your main.rs after the call to WindowManager::grab_keys_and_run and all internal state will still be accessible (though methods requiring the use of the XConn will fail.

pub fn detect_screens(&mut self) -> Result<()>

Query the XConn for the current connected Screen list and reposition displayed Workspace instances if needed.

pub fn conn(&self) -> &X

Get an immutable reference to the underlying XConn impl that backs this WindowManager

pub fn conn_mut(&mut self) -> &mut X

Get an mutable reference to the underlying XConn impl that backs this WindowManager

pub fn log(&self, msg: &str) -> Result<()>

Log information out at INFO level for picking up by external programs

Example

use penrose::{
    core::{Config, WindowManager},
    xcb::XcbConnection,
    logging_error_handler
};

let mut wm = WindowManager::new(
    Config::default(),
    XcbConnection::new().unwrap(),
    vec![],
    logging_error_handler(),
);

wm.log("hello from penrose").unwrap();

pub fn cycle_screen(&mut self, direction: Direction) -> Result<()>

Cycle between known screens. Does not wrap from first to last

Example

use penrose::{
    core::{Config, WindowManager},
    xcb::XcbConnection,
    Forward, logging_error_handler
};

let mut wm = WindowManager::new(
    Config::default(),
    XcbConnection::new().unwrap(),
    vec![],
    logging_error_handler(),
);

if let Err(e) = wm.init() {
    panic!("failed to initialise WindowManager: {}", e);
}

assert_eq!(wm.active_screen_index(), 0);
wm.cycle_screen(Forward).unwrap();
assert_eq!(wm.active_screen_index(), 1);

pub fn cycle_workspace(&mut self, direction: Direction) -> Result<()>

Cycle between workspaces on the current screen. This will pull workspaces to the screen if they are currently displayed on another screen.

pub fn drag_workspace(&mut self, direction: Direction) -> Result<()>

Move the currently focused workspace to the next screen in 'direction'

pub fn cycle_client(&mut self, direction: Direction) -> Result<()>

Cycle between clients for the active workspace

pub fn focus_client(
&mut self,
selector: &Selector<'_, Client>
) -> Result<WinId, Option<WinId>>

Focus the Client matching the given Selector

Errors

If the selector matches a known client then that client is focused and Ok(id) is returned. If the selector doesn't match (either it was invalid or there is no focused client) then Err(self.focused_client_id()) is returned.

pub fn rotate_clients(&mut self, direction: Direction) -> Result<()>

Rotate the client stack on the active workspace

pub fn drag_client(&mut self, direction: Direction) -> Result<()>

Move the focused client through the stack of clients on the active workspace

pub fn cycle_layout(&mut self, direction: Direction) -> Result<()>

Cycle between layouts for the active workspace

pub fn update_max_main(&mut self, change: Change) -> Result<()>

Increase or decrease the number of clients in the main area by 1

pub fn update_main_ratio(&mut self, change: Change) -> Result<()>

Increase or decrease the current layout main_ratio by main_ratio_step

pub fn exit(&mut self) -> Result<()>

Shut down the WindowManager, running any required cleanup and exiting penrose

pub fn current_layout_symbol(&self) -> &str

The layout symbol for the layout currently being used on the active workspace

pub fn set_root_window_name(&self, s: &str) -> Result<()>

Set the root X window name. Useful for exposing information to external programs

pub fn set_client_insert_point(&mut self, cip: InsertPoint) -> Result<()>

Set the insert point for new clients. Default is to insert at index 0.

pub fn focus_workspace(
&mut self,
selector: &Selector<'_, Workspace>
) -> Result<()>

Set the displayed workspace for the focused screen to be index in the list of workspaces passed at init. This will panic if the index passed is out of bounds which is only possible if you manually bind an action to this with an invalid index. You should almost always be using the gen_keybindings! macro to set up your keybindings so this is not normally an issue.

pub fn toggle_workspace(&mut self) -> Result<()>

Switch focus back to the last workspace that had focus.

pub fn client_to_workspace(
&mut self,
selector: &Selector<'_, Workspace>
) -> Result<()>

Move the focused client to the workspace matching 'selector'.

pub fn client_to_screen(
&mut self,
selector: &Selector<'_, Screen>
) -> Result<()>

Move the focused client to the active workspace on the screen matching 'selector'.

pub fn toggle_client_fullscreen(
&mut self,
selector: &Selector<'_, Client>
) -> Result<()>

Toggle the fullscreen state of the given client ID

pub fn kill_client(&mut self) -> Result<()>

Kill the focused client window.

pub fn screen(&self, selector: &Selector<'_, Screen>) -> Option<&Screen>

Get a reference to the first Screen satisfying 'selector'. WinId selectors will return the screen containing that Client if the client is known. NOTE: It is not possible to get a mutable reference to a Screen.

pub fn focused_workspaces(&self) -> Vec<usize>

The currently focused workspace indices being shown on each screen

pub fn add_workspace(&mut self, index: usize, ws: Workspace) -> Result<()>

Add a new workspace at index, shifting all workspaces with indices greater to the right.

pub fn push_workspace(&mut self, ws: Workspace) -> Result<()>

Add a new workspace at the end of the current workspace list

pub fn remove_workspace(
&mut self,
selector: &Selector<'_, Workspace>
) -> Result<Option<Workspace>>

Remove a Workspace from the WindowManager. All clients that were present on the removed workspace will be destroyed. WinId selectors will be ignored.

pub fn workspace(
&self,
selector: &Selector<'_, Workspace>
) -> Option<&Workspace>

Get a reference to the first Workspace satisfying 'selector'. WinId selectors will return the workspace containing that Client if the client is known.

pub fn workspace_mut(
&mut self,
selector: &Selector<'_, Workspace>
) -> Option<&mut Workspace>

Get a mutable reference to the first Workspace satisfying 'selector'. WinId selectors will return the workspace containing that Client if the client is known.

pub fn all_workspaces(
&self,
selector: &Selector<'_, Workspace>
) -> Vec<&Workspace>

Get a vector of references to Workspaces satisfying 'selector'. WinId selectors will return a vector with the workspace containing that Client if the client is known. Otherwise an empty vector will be returned.

pub fn all_workspaces_mut(
&mut self,
selector: &Selector<'_, Workspace>
) -> Vec<&mut Workspace>

Get a vector of mutable references to Workspaces satisfying 'selector'. WinId selectors will return a vector with the workspace containing that Client if the client is known. Otherwise an empty vector will be returned.

pub fn set_workspace_name(
&mut self,
name: impl Into<String>,
selector: Selector<'_, Workspace>
) -> Result<()>

Set the name of the selected Workspace

pub fn client(&self, selector: &Selector<'_, Client>) -> Option<&Client>

Take a reference to the first Client found matching 'selector'

pub fn client_mut(
&mut self,
selector: &Selector<'_, Client>
) -> Option<&mut Client>

Take a mutable reference to the first Client found matching 'selector'

pub fn all_clients(&self, selector: &Selector<'_, Client>) -> Vec<&Client>

Get a vector of references to the Clients found matching 'selector'. The resulting vector is sorted by Client id.

pub fn all_clients_mut(
&mut self,
selector: &Selector<'_, Client>
) -> Vec<&mut Client>

Get a vector of mutable references to the Clients found matching 'selector'. The resulting vector is sorted by Client id.

pub fn n_screens(&self) -> usize

The number of detected screens currently being tracked by the WindowManager.

pub fn screen_size(&self, screen_index: usize) -> Option<Region>

The current effective screen size of the target screen. Effective screen size is the physical screen size minus any space reserved for a status bar.

pub fn position_client(
&self,
id: WinId,
region: Region,
stack_above: bool
) -> Result<()>

Position an individual client on the display. (x,y) coordinates are absolute (i.e. relative to the root window not any individual screen).

pub fn show_client(&mut self, id: WinId) -> Result<()>

Make the Client with ID 'id' visible at its last known position.

pub fn hide_client(&mut self, id: WinId) -> Result<()>

Hide the Client with ID 'id'.

pub fn layout_screen(&mut self, screen_index: usize) -> Result<()>

Layout the workspace currently shown on the given screen index.

pub fn active_screen_index(&self) -> usize

An index into the WindowManager known screens for the screen that is currently focused

impl WindowManager<XcbConnection>

pub fn xcb_connectction(&self) -> &Connection

Get a handle on the underlying [XCB Connection][::xcb::Connection] used by Api to communicate with the X server.

pub fn known_atoms(&self) -> &HashMap<Atom, u32>

The current interned Atom values known to the underlying XcbConnection