Struct penrose::core::Config

pub struct Config<X>
where
    X: XConn,
{
    pub normal_border: Color,
    pub focused_border: Color,
    pub border_width: u32,
    pub focus_follow_mouse: bool,
    pub default_layouts: LayoutStack,
    pub tags: Vec<String>,
    pub floating_classes: Vec<String>,
    pub startup_hook: Option<Box<dyn StateHook<X>>>,
    pub event_hook: Option<Box<dyn EventHook<X>>>,
    pub manage_hook: Option<Box<dyn ManageHook<X>>>,
    pub refresh_hook: Option<Box<dyn StateHook<X>>>,
    pub layout_hook: Option<Box<dyn LayoutHook<X>>>,
}

The user specified config options for how the window manager should run

Fields
normal_border: Color

The RGBA color to use for normal (unfocused) window borders

focused_border: Color

The RGBA color to use for the focused window border

border_width: u32

The width in pixels to use for drawing window borders

focus_follow_mouse: bool

Whether or not the mouse entering a new window should set focus

default_layouts: LayoutStack

The stack of layouts to use for each workspace

tags: Vec<String>

The ordered set of workspace tags to use on window manager startup

floating_classes: Vec<String>

Window classes that should always be assigned floating positions rather than tiled

startup_hook: Option<Box<dyn StateHook<X>>>

A StateHook to run before entering the main event loop

event_hook: Option<Box<dyn EventHook<X>>>

A StateHook to run before processing each XEvent

manage_hook: Option<Box<dyn ManageHook<X>>>

A ManageHook to run after each new window becomes managed by the window manager

refresh_hook: Option<Box<dyn StateHook<X>>>

A StateHook to run every time the on screen X state is refreshed

layout_hook: Option<Box<dyn LayoutHook<X>>>

A LayoutHook to run when positioning clients on the screen

Implementations

impl<X> Config<X>
where X: XConn,

pub fn compose_or_set_startup_hook<H>(&mut self, hook: H)
where H: StateHook<X> + 'static, X: 'static,

Set the startup_hook or compose it with what is already set.

The new hook will run before what was there before.

pub fn compose_or_set_event_hook<H>(&mut self, hook: H)
where H: EventHook<X> + 'static, X: 'static,

Set the event_hook or compose it with what is already set.

The new hook will run before what was there before.

pub fn compose_or_set_manage_hook<H>(&mut self, hook: H)
where H: ManageHook<X> + 'static, X: 'static,

Set the manage_hook or compose it with what is already set.

The new hook will run before what was there before.

pub fn compose_or_set_refresh_hook<H>(&mut self, hook: H)
where H: StateHook<X> + 'static, X: 'static,

Set the refresh_hook or compose it with what is already set.

The new hook will run before what was there before.

pub fn compose_or_set_layout_hook<H>(&mut self, hook: H)
where H: LayoutHook<X> + 'static, X: 'static,

Set the layout_hook or compose it with what is already set.

The new hook will run before what was there before.