Struct penrose::core::WindowManager

pub struct WindowManager<X>
where
    X: XConn,
{
    pub state: State<X>,
    /* private fields */
}

A top level struct holding all of the state required to run as an X11 window manager.

This allows for final configuration to be carried out before entering the main event loop.

Fields
state: State<X>

The mutable State of the window manager

Implementations

impl<X> WindowManager<X>
where X: XConn,

pub fn new( config: Config<X>, key_bindings: KeyBindings<X>, mouse_bindings: MouseBindings<X>, x: X ) -> Result<Self>

Construct a new WindowManager with the provided config and X connection.

If you need to set State extensions, call WindowManager::add_extension after constructing your initial WindowManager.

pub fn add_extension<E: Any>(&mut self, extension: E)

Add a typed State extension to this WindowManager.

pub fn run(self) -> Result<()>

Start the WindowManager and run it until told to exit.

Any provided startup hooks will be run after setting signal handlers and grabbing key / mouse bindings from the X server. Any set up you need to do should be run explicitly before calling this method or as part of a startup hook.

Existing clients

An attempt will be made to pull any existing clients already present into the current WindowManager state. This is done on a “best effort” basis to manage existing clients on the workspaces they were present on previously. If you are planning on making use of this functionality for more than recovering from a crash it is advised that you add EWMH hooks to your Config so that there is more information available to correctly position your existing clients. Startup hooks are run before this takes place so that there is an opportunity to handle restoring any state being held outside of the main WindowManager data structures.

NOTE: This is not guaranteed to preserve the stacking order or correctly handle any clients that were on invisible workspaces / workspaces that no longer exist and that the workspace containing the previously active client will be placed on the first available screen.