Struct penrose::core::Workspace

pub struct Workspace { /* fields omitted */ }

A Workspace represents a named set of clients that are tiled according to a specific layout. Layout properties are tracked per workspace and clients are referenced by ID. Workspaces are independant of monitors and can be moved between monitors freely, bringing their clients with them.

The parent WindowManager struct tracks which client is focused from the point of view of the X server by checking focus at the Workspace level whenever a new Workspace becomes active.

Implementations

impl Workspace

pub fn new(name: impl Into<String>, layouts: Vec<Layout>) -> Self

Construct a new Workspace with the given name and choice of Layouts

pub fn name(&self) -> &str

The name of this workspace

pub fn len(&self) -> usize

The number of clients currently on this workspace

pub fn is_empty(&self) -> bool

Is this Workspace currently empty?

pub fn iter(&self) -> Iter<'_, WinId>

Iterate over the clients on this workspace in position order

pub fn iter_mut(&mut self) -> IterMut<'_, WinId>

Iterate over the clients on this workspace in position order

pub fn focused_client(&self) -> Option<WinId>

A reference to the currently focused client if there is one

pub fn add_client(&mut self, id: WinId, ip: &InsertPoint)

Add a new client to this workspace at the top of the stack and focus it

pub fn focus_client(&mut self, id: WinId) -> Option<WinId>

Focus the client with the given id, returns an option of the previously focused client if there was one

pub fn remove_client(&mut self, id: WinId) -> Option<WinId>

Remove a target client, retaining focus at the same position in the stack. Returns the removed client if there was one to remove.

pub fn remove_focused_client(&mut self) -> Option<WinId>

Remove the currently focused client, keeping focus at the same position in the stack. Returns the removed client if there was one to remove.

pub fn try_set_layout(&mut self, symbol: &str) -> Option<&Layout>

Set the active layout by symbol name if it is available. Returns a reference to active layout if it was able to be set.

pub fn cycle_layout(&mut self, direction: Direction) -> &str

Cycle through the available layouts on this workspace

pub fn layout_symbol(&self) -> &str

The symbol of the currently used layout (passed on creation)

pub fn layout_conf(&self) -> LayoutConf

The LayoutConf of the currently active Layout. Used by the WindowManager to determine when and how the layout function should be applied.

pub fn cycle_client(&mut self, direction: Direction) -> Option<(WinId, WinId)>

Cycle focus through the clients on this workspace

pub fn drag_client(&mut self, direction: Direction) -> Option<WinId>

Drag the focused client through the stack, retaining focus

pub fn rotate_clients(&mut self, direction: Direction)

Rotate the client stack in the given direction

pub fn update_max_main(&mut self, change: Change)

Increase or decrease the number of possible clients in the main area of the current Layout

pub fn update_main_ratio(&mut self, change: Change, step: f32)

Increase or decrease the size of the main area for the current Layout