Type Alias penrose::core::ClientSpace

pub type ClientSpace = Workspace<Xid>;

The pure client state information for a single Workspace

Aliased Type

struct ClientSpace { /* private fields */ }

Implementations

impl<T> Workspace<T>

pub fn new<S>( id: usize, tag: S, layouts: LayoutStack, stack: Option<Stack<T>> ) -> Selfwhere S: Into<String>,

Create a new Workspace with the given layouts and stack.

pub fn id(&self) -> usize

A fixed integer ID for this workspace.

pub fn tag(&self) -> &str

The string tag for this workspace.

pub fn layout_name(&self) -> String

The name of the currently active layout being used by this workspace

pub fn is_empty(&self) -> bool

Whether or not this workspace currently holds any windows

pub fn focus(&self) -> Option<&T>

An immutable reference to the focused window for this workspace if there is one

pub fn clients(&self) -> impl Iterator<Item = &T>

An iterator over all windows in this workspace.

pub fn handle_message<M>(&mut self, m: M)where M: IntoMessage,

Pass the given message on to the currently focused layout.

pub fn broadcast_message<M>(&mut self, m: M)where M: IntoMessage,

Pass the given message on to all layouts available to this workspace.

pub fn next_layout(&mut self)

Switch to the next available layout for this workspace.

pub fn previous_layout(&mut self)

Switch to the previous available layout for this workspace.

pub fn set_available_layouts(&mut self, layouts: LayoutStack) -> LayoutStack

Replace the current LayoutStack with a new one, returning the layouts that were previously active.

pub fn set_layout_by_name(&mut self, name: &str)

Attempt to set the active Layout by name if it is available.

Note that some layouts have a dynamically set name and this method will fail to locate such a layout if the current name does not match what you have provided.

impl<T: PartialEq> Workspace<T>

pub fn contains(&self, t: &T) -> bool

Check if a given window is currently part of this workspace