Module penrose::core::xconnection

An abstraciton layer for talking to an underlying X server.

An implementation of the XConn trait is required for running a WindowManager. The choice of back end (e.g. xlib, xcb...) is an implementation detail that does not surface in the WindowManager itself. All low level details of working with the X server should be captured in this trait, though accessing backend specific functionality is possible by writing an impl block for WindowManager<YourXConn> if desired.

Structs

AtomIter

MockXConn
A dummy XConn implementation for testing

Enums

Atom
A Penrose internal representation of X atoms.

XEvent
Wrapper around the low level X event types that correspond to request / response data when communicating with the X server itself.

Traits

StubXConn
A really simple stub implementation of XConn to simplify setting up test cases.

XConn
A handle on a running X11 connection that we can use for issuing X requests.