Enum penrose::core::xconnection::XEvent

pub enum XEvent {
    MouseEvent(MouseEvent),
    KeyPress(KeyCode),
    MapRequest {
        id: WinId,
        ignore: bool,
    },
    Enter {
        id: WinId,
        rpt: Point,
        wpt: Point,
    },
    Leave {
        id: WinId,
        rpt: Point,
        wpt: Point,
    },
    Destroy {
        id: WinId,
    },
    ScreenChange,
    RandrNotify,
    ConfigureNotify {
        id: WinId,
        r: Region,
        is_root: bool,
    },
    PropertyNotify {
        id: WinId,
        atom: String,
        is_root: bool,
    },
    ClientMessage {
        id: WinId,
        dtype: String,
        data: Vec<usize>,
    },
}

Wrapper around the low level X event types that correspond to request / response data when communicating with the X server itself.

The variant names and data have developed with the reference xcb implementation in mind but should be applicable for all back ends.

Variants

MouseEvent(MouseEvent)

The mouse has moved or a mouse button has been pressed

KeyPress(KeyCode)

A grabbed key combination has been entered by the user

MapRequest

A client window is requesting to be positioned and rendered on the screen

Fields of MapRequest

id: WinId

The ID of the window that wants to be mapped

ignore: bool

Whether or not the WindowManager should handle this window.

Enter

The mouse pointer has entered a new client window

Fields of Enter

id: WinId

The ID of the window that was entered

rpt: Point

Absolute coordinate of the event

wpt: Point

Coordinate of the event relative to top-left of the window itself

Leave

The mouse pointer has left the current client window

Fields of Leave

id: WinId

The ID of the window that was left

rpt: Point

Absolute coordinate of the event

wpt: Point

Coordinate of the event relative to top-left of the window itself

Destroy

A client window has been closed

Fields of Destroy

id: WinId

The ID of the window being destroyed

ScreenChange

Focus has moved to a different screen

RandrNotify

A randr action has occured (new outputs, resolution change etc)

ConfigureNotify

Client config has changed in some way

Fields of ConfigureNotify

id: WinId

The ID of the window that had a property changed

r: Region

The new window size

is_root: bool

Is this window the root window?

PropertyNotify

A client property has changed in some way

Fields of PropertyNotify

id: WinId

The ID of the window that had a property changed

atom: String

The property that changed

is_root: bool

Is this window the root window?

ClientMessage

A message has been sent to a particular client

Fields of ClientMessage

id: WinId

The ID of the window that sent the message

dtype: String

The data type being set

data: Vec<usize>

The data itself