Trait penrose::core::xconnection::StubXConn

pub trait StubXConn {
    fn mock_hydrate(&mut self) -> Result<()> { ... }

    fn mock_flush(&self) -> bool { ... }

    fn mock_wait_for_event(&self) -> Result<XEvent> { ... }

    fn mock_current_outputs(&self) -> Vec<Screen> { ... }

    fn mock_cursor_position(&self) -> Point { ... }

    fn mock_send_client_event(&self, _: WinId, _: &str) -> Result<()> { ... }

    fn mock_focused_client(&self) -> WinId { ... }

    fn mock_window_should_float(&self, _: WinId, _: &[&str]) -> bool { ... }

    fn mock_is_managed_window(&self, _: WinId) -> bool { ... }

    fn mock_window_geometry(&self, _: WinId) -> Result<Region> { ... }

    fn mock_query_for_active_windows(&self) -> Vec<WinId> { ... }

    fn mock_str_prop(&self, _: u32, name: &str) -> Result<String> { ... }

    fn mock_atom_prop(&self, id: u32, _: &str) -> Result<u32> { ... }

    fn mock_intern_atom(&self, _: &str) -> Result<u32> { ... }

    fn mock_warp_cursor(&self, _: Option<WinId>, _: &Screen) { ... }

    fn mock_focus_client(&self, _: WinId) { ... }

    fn mock_position_window(&self, _: WinId, _: Region, _: u32, _: bool) { ... }

    fn mock_raise_window(&self, _: WinId) { ... }

    fn mock_mark_new_window(&self, _: WinId) { ... }

    fn mock_map_window(&self, _: WinId) { ... }

    fn mock_unmap_window(&self, _: WinId) { ... }

    fn mock_set_client_border_color(&self, _: WinId, _: u32) { ... }

    fn mock_grab_keys(&self, _: &KeyBindings<Self>, _: &MouseBindings<Self>)
    where
        Self: Sized,
    { ... }

    fn mock_set_wm_properties(&self, _: &[&str]) { ... }

    fn mock_update_desktops(&self, _: &[&str]) { ... }

    fn mock_update_known_clients(&self, _: &[WinId]) { ... }

    fn mock_set_current_workspace(&self, _: usize) { ... }

    fn mock_set_root_window_name(&self, _: &str) { ... }

    fn mock_set_client_workspace(&self, _: WinId, _: usize) { ... }

    fn mock_toggle_client_fullscreen(&self, _: WinId, _: bool) { ... }

    fn mock_cleanup(&self) { ... }
}

A really simple stub implementation of XConn to simplify setting up test cases.

Intended use is to override the mock_* methods that you need for running your test case in order to inject behaviour into a WindowManager instance which is driven by X server state. StubXConn will then implement XConn and call through to your overwritten methods or the provided default.

This is being done to avoid providing broken default methods on the real XConn trait that would make writing real impls more error prone if and when new methods are added to the trait.

Provided methods

fn mock_hydrate(&mut self) -> Result<()>

Mocked version of hydrate

fn mock_flush(&self) -> bool

Mocked version of flush

fn mock_wait_for_event(&self) -> Result<XEvent>

Mocked version of wait_for_event

fn mock_current_outputs(&self) -> Vec<Screen>

Mocked version of current_outputs

fn mock_cursor_position(&self) -> Point

Mocked version of cursor_position

fn mock_send_client_event(&self, _: WinId, _: &str) -> Result<()>

Mocked version of send_client_event

fn mock_focused_client(&self) -> WinId

Mocked version of focused_client

fn mock_window_should_float(&self, _: WinId, _: &[&str]) -> bool

Mocked version of window_should_float

fn mock_is_managed_window(&self, _: WinId) -> bool

Mocked version of is_managed_window

fn mock_window_geometry(&self, _: WinId) -> Result<Region>

Mocked version of window_geometry

fn mock_query_for_active_windows(&self) -> Vec<WinId>

Mocked version of query_for_active_windows

fn mock_str_prop(&self, _: u32, name: &str) -> Result<String>

Mocked version of str_prop

fn mock_atom_prop(&self, id: u32, _: &str) -> Result<u32>

Mocked version of atom_prop

fn mock_intern_atom(&self, _: &str) -> Result<u32>

Mocked version of intern_atom

fn mock_warp_cursor(&self, _: Option<WinId>, _: &Screen)

Mocked version of warp_cursor

fn mock_focus_client(&self, _: WinId)

Mocked version of focus_client

fn mock_position_window(&self, _: WinId, _: Region, _: u32, _: bool)

Mocked version of position_window

fn mock_raise_window(&self, _: WinId)

Mocked version of raise_window

fn mock_mark_new_window(&self, _: WinId)

Mocked version of mark_new_window

fn mock_map_window(&self, _: WinId)

Mocked version of map_window

fn mock_unmap_window(&self, _: WinId)

Mocked version of unmap_window

fn mock_set_client_border_color(&self, _: WinId, _: u32)

Mocked version of set_client_border_color

fn mock_grab_keys(&self, _: &KeyBindings<Self>, _: &MouseBindings<Self>) where
Self: Sized,

Mocked version of grab_keys

fn mock_set_wm_properties(&self, _: &[&str])

Mocked version of set_wm_properties

fn mock_update_desktops(&self, _: &[&str])

Mocked version of update_desktops

fn mock_update_known_clients(&self, _: &[WinId])

Mocked version of update_known_clients

fn mock_set_current_workspace(&self, _: usize)

Mocked version of set_current_workspace

fn mock_set_root_window_name(&self, _: &str)

Mocked version of set_root_window_name

fn mock_set_client_workspace(&self, _: WinId, _: usize)

Mocked version of set_client_workspace

fn mock_toggle_client_fullscreen(&self, _: WinId, _: bool)

Mocked version of toggle_client_fullscreen

fn mock_cleanup(&self)

Mocked version of cleanup

Loading content...