Trait penrose::core::xconnection::XConn

pub trait XConn {
    fn hydrate(&mut self) -> Result<()>;

    fn flush(&self) -> bool;

    fn wait_for_event(&self) -> Result<XEvent>;

    fn current_outputs(&self) -> Vec<Screen>;

    fn cursor_position(&self) -> Point;

    fn position_window(
        &self, 
        id: WinId, 
        r: Region, 
        border: u32, 
        stack_above: bool
    );

    fn raise_window(&self, id: WinId);

    fn mark_new_window(&self, id: WinId);

    fn map_window(&self, id: WinId);

    fn unmap_window(&self, id: WinId);

    fn send_client_event(&self, id: WinId, atom_name: &str) -> Result<()>;

    fn focused_client(&self) -> WinId;

    fn focus_client(&self, id: WinId);

    fn set_client_border_color(&self, id: WinId, color: u32);

    fn grab_keys(
        &self, 
        key_bindings: &KeyBindings<Self>, 
        mouse_bindings: &MouseBindings<Self>
    )
    where
        Self: Sized;

    fn set_wm_properties(&self, workspaces: &[&str]);

    fn update_desktops(&self, workspaces: &[&str]);

    fn update_known_clients(&self, clients: &[WinId]);

    fn set_current_workspace(&self, wix: usize);

    fn set_root_window_name(&self, name: &str);

    fn set_client_workspace(&self, id: WinId, wix: usize);

    fn toggle_client_fullscreen(&self, id: WinId, client_is_fullscreen: bool);

    fn window_should_float(&self, id: WinId, floating_classes: &[&str]) -> bool;

    fn is_managed_window(&self, id: WinId) -> bool;

    fn window_geometry(&self, id: WinId) -> Result<Region>;

    fn warp_cursor(&self, id: Option<WinId>, screen: &Screen);

    fn query_for_active_windows(&self) -> Vec<WinId>;

    fn str_prop(&self, id: u32, name: &str) -> Result<String>;

    fn atom_prop(&self, id: u32, name: &str) -> Result<u32>;

    fn intern_atom(&self, atom: &str) -> Result<u32>;

    fn cleanup(&self);
}

A handle on a running X11 connection that we can use for issuing X requests.

XConn is intended as an abstraction layer to allow for communication with the underlying display system (assumed to be X) using whatever mechanism the implementer wishes. In theory, it should be possible to write an implementation that allows penrose to run on systems not using X as the windowing system but X idioms and high level event types / client interations are assumed.

Required methods

fn hydrate(&mut self) -> Result<()>

Hydrate this XConn to restore internal state following serde deserialization

fn flush(&self) -> bool

Flush pending actions to the X event loop

fn wait_for_event(&self) -> Result<XEvent>

Wait for the next event from the X server and return it as an XEvent

fn current_outputs(&self) -> Vec<Screen>

Determine the currently connected CRTCs and return their details

fn cursor_position(&self) -> Point

Determine the current (x,y) position of the cursor relative to the root window.

fn position_window(&self, id: WinId, r: Region, border: u32, stack_above: bool)

Reposition the window identified by 'id' to the specifed region

fn raise_window(&self, id: WinId)

Raise the window to the top of the stack so it renders above peers

fn mark_new_window(&self, id: WinId)

Mark the given window as newly created

fn map_window(&self, id: WinId)

Map a window to the display. Called each time a map_notify event is received

fn unmap_window(&self, id: WinId)

Unmap a window from the display. Called each time an unmap_notify event is received

fn send_client_event(&self, id: WinId, atom_name: &str) -> Result<()>

Send an X event to the target window

fn focused_client(&self) -> WinId

Return the client ID of the crate::core::client::Client that currently holds X focus

fn focus_client(&self, id: WinId)

Mark the given crate::core::client::Client as having focus

fn set_client_border_color(&self, id: WinId, color: u32)

Change the border color for the given client

fn grab_keys(
&self,
key_bindings: &KeyBindings<Self>,
mouse_bindings: &MouseBindings<Self>
) where
Self: Sized,

Notify the X server that we are intercepting the user specified key bindings and prevent them being passed through to the underlying applications. This is what determines which key press events end up being sent through in the main event loop for the WindowManager.

fn set_wm_properties(&self, workspaces: &[&str])

Set required EWMH properties to ensure compatability with external programs

fn update_desktops(&self, workspaces: &[&str])

Update the root window properties with the current desktop details

fn update_known_clients(&self, clients: &[WinId])

Update the root window properties with the current client details

fn set_current_workspace(&self, wix: usize)

Update which desktop is currently focused

fn set_root_window_name(&self, name: &str)

Set the WM_NAME prop of the root window

fn set_client_workspace(&self, id: WinId, wix: usize)

Update which desktop a client is currently on

fn toggle_client_fullscreen(&self, id: WinId, client_is_fullscreen: bool)

Toggle the fullscreen state of the given client ID with the X server

fn window_should_float(&self, id: WinId, floating_classes: &[&str]) -> bool

Determine whether the target window should be tiled or allowed to float

fn is_managed_window(&self, id: WinId) -> bool

Check to see if this window is one that we should be handling or not

fn window_geometry(&self, id: WinId) -> Result<Region>

Return the current (x, y, w, h) dimensions of the requested window

fn warp_cursor(&self, id: Option<WinId>, screen: &Screen)

Warp the cursor to be within the specified window. If id == None then behaviour is definined by the implementor (e.g. warp cursor to active window, warp to center of screen)

fn query_for_active_windows(&self) -> Vec<WinId>

Run on startup/restart to determine already running windows that we need to track

fn str_prop(&self, id: u32, name: &str) -> Result<String>

Query a string property for a window by window ID and poperty name. Can fail if the property name is invalid or we get a malformed response from xcb.

fn atom_prop(&self, id: u32, name: &str) -> Result<u32>

Fetch an atom prop by name for a particular window ID Can fail if the property name is invalid or we get a malformed response from xcb.

fn intern_atom(&self, atom: &str) -> Result<u32>

Intern an X atom by name and return the corresponding ID

fn cleanup(&self)

Perform any state cleanup required prior to shutting down the window manager

Loading content...