Module penrose::draw::bar

Simple text based status bars

This module provides a framework for writing simple, text based status bars such as those seen in dwm and xmonad. A StatusBar itself acts as a multiplexer for the contained widgets and each of their Hook triggers, deferring to each individual HookableWidget for how it should be rendered to the screen.

A minimal example bar configuration is provided in the form of dwm_bar which aims to emulate the behaviour and appearance of the built in status bar from dwm.

Structs

StatusBar
A simple status bar that works via hooks

Enums

Position
The position of a status bar

Functions

dwm_bar
Create a default dwm style status bar that displays content pulled from the WM_NAME property of the root window.