Struct penrose::draw::LinesWithSelection

pub struct LinesWithSelection { /* fields omitted */ }

A set of lines that highlights the currently selected line.

Implementations

impl LinesWithSelection

pub fn new(
font: String,
point_size: i32,
padding: f64,
bg: Color,
fg: Color,
bg_sel: Color,
fg_sel: Color,
n_lines: usize,
greedy: bool
) -> Self

Construct a new LinesWithSelection

pub fn set_input(&mut self, lines: Vec<String>) -> Result<()>

Set the displayed lines and selected index.

pub fn set_selected(&mut self, selected: usize) -> Result<()>

Set the currently selected index

Errors

Fails if the provided index is out of bounds

pub fn set_n_lines(&mut self, n_lines: usize)

Set the maximum number of lines from the input that will be displayed.

Defaults to 10

pub fn selected(&self) -> Option<&str>

The currently selected line (if there is one)

pub fn selected_index(&self) -> usize

The currently selected index