Trait penrose::draw::Draw

pub trait Draw {
    type Ctx: DrawContext;
    fn new_window(
        &mut self, 
        ty: WinType, 
        r: Region, 
        managed: bool
    ) -> Result<WinId>;

    fn screen_sizes(&self) -> Result<Vec<Region>>;

    fn register_font(&mut self, font_name: &str);

    fn context_for(&self, id: WinId) -> Result<Self::Ctx>;

    fn temp_context(&self, w: u32, h: u32) -> Result<Self::Ctx>;

    fn flush(&self, id: WinId);

    fn map_window(&self, id: WinId);

    fn unmap_window(&self, id: WinId);

    fn destroy_window(&mut self, id: WinId);

    fn replace_prop(&self, id: WinId, prop: Atom, val: PropVal<'_>);
}

A simple drawing abstraction

Draw is not intended for use in writing full GUI interfaces, rather it is a simple abstraction layer to allow for the creation of minimal UI elements such as status bars, menus and dialogs. Each Draw should also provide an acompanying DrawContext that impl that is used by consumers (such as the status bar) for actually drawing to the screen, which the parent Draw is responsible for resource management and mapping / unmapping the created windows.

Associated Types

type Ctx: DrawContext

The type of drawing context used for drawing

Loading content...
Required methods

fn new_window(&mut self, ty: WinType, r: Region, managed: bool) -> Result<WinId>

Create a new client window with a canvas for drawing

fn screen_sizes(&self) -> Result<Vec<Region>>

Get the size of the target screen in pixels

fn register_font(&mut self, font_name: &str)

Register a font by name for later use

fn context_for(&self, id: WinId) -> Result<Self::Ctx>

Get a new DrawContext for the target window

fn temp_context(&self, w: u32, h: u32) -> Result<Self::Ctx>

Get a new temporary DrawContext that will be destroyed when dropped

fn flush(&self, id: WinId)

Flush pending actions

fn map_window(&self, id: WinId)

Map the target window to the screen

fn unmap_window(&self, id: WinId)

Unmap the target window from the screen

fn destroy_window(&mut self, id: WinId)

Destroy the target window

fn replace_prop(&self, id: WinId, prop: Atom, val: PropVal<'_>)

Replace a property value on a window.

See the documentation for the C level XCB API for the correct property type for each prop.

Loading content...