Trait penrose::draw::DrawContext

pub trait DrawContext {
    fn font(&mut self, font_name: &str, point_size: i32) -> Result<()>;

    fn color(&mut self, color: &Color);

    fn clear(&mut self);

    fn translate(&self, dx: f64, dy: f64);

    fn set_x_offset(&self, x: f64);

    fn set_y_offset(&self, y: f64);

    fn rectangle(&self, x: f64, y: f64, w: f64, h: f64);

    fn text(
        &self, 
        s: &str, 
        h_offset: f64, 
        padding: (f64, f64)
    ) -> Result<(f64, f64)>;

    fn text_extent(&self, s: &str) -> Result<(f64, f64)>;

    fn flush(&self);
}

Used for simple drawing to the screen

Required methods

fn font(&mut self, font_name: &str, point_size: i32) -> Result<()>

Set the active font, must have been registered on the partent Draw

fn color(&mut self, color: &Color)

Set the color used for subsequent drawing operations

fn clear(&mut self)

Clears the context

fn translate(&self, dx: f64, dy: f64)

Translate this context by (dx, dy) from its current position

fn set_x_offset(&self, x: f64)

Set the x offset for this context absolutely

fn set_y_offset(&self, y: f64)

Set the y offset for this context absolutely

fn rectangle(&self, x: f64, y: f64, w: f64, h: f64)

Draw a filled rectangle using the current color

fn text(
&self,
s: &str,
h_offset: f64,
padding: (f64, f64)
) -> Result<(f64, f64)>

Render 's' using the current font with the supplied padding. returns the extent taken up by the rendered text

fn text_extent(&self, s: &str) -> Result<(f64, f64)>

Determine the pixel width of a given piece of text using the current font

fn flush(&self)

Flush pending actions

Loading content...