Module penrose::draw::widget

Self rendering building blocks

The widgets defined in this module are backed by the Draw trait and are composable to build up more complex layouts and UI elements.

Modules

bar
Widgets intended for use in statusbars

base
Base widgets for building more complex structures

Structs

ActiveWindowName
A text widget that is set via updating the root window name a la dwm

CurrentLayout
A simple widget that displays the active layout symbol

InputBox
A simple text box that can be driven by user keyboard input

LinesWithSelection
A set of lines that highlights the currently selected line.

RootWindowName
A text widget that is set via updating the root window name a la dwm

Text
A simple piece of static text with an optional background color.

Workspaces
A simple workspace indicator for a status bar

Traits

HookableWidget
A status bar Widget that can be automatically rendered using a DrawContext when triggered via WindowManager Hook calls.

KeyboardControlled
Something that can respond to user KeyPress events

Widget
A status bar widget that can be rendered using a DrawContext