Struct penrose::draw::widget::base::Text

pub struct Text { /* fields omitted */ }

A simple piece of static text with an optional background color.

Can be used as a simple static element in a status bar or as an inner element for rendering more complex text based widgets.

Implementations

impl Text

pub fn new(
txt: impl Into<String>,
style: &TextStyle,
is_greedy: bool,
right_justified: bool
) -> Self

Construct a new Text

pub fn get_text(&self) -> &String

Borrow the current contents of the widget.

pub fn get_text_mut(&mut self) -> &mut String

Mutably borrow the current contents of the widget.

pub fn set_text(&mut self, txt: impl Into<String>)

Set the rendered text and trigger a redraw

pub fn force_draw(&mut self)

Force this text widget to redraw on the next render request. Mostly used when being wrapped by another widget.