Function penrose::extensions::actions::create_or_switch_to_workspace

pub fn create_or_switch_to_workspace<X>(
    get_name: fn() -> Option<String>,
    layouts: LayoutStack
) -> Box<dyn KeyEventHandler<X>>
where
    X: XConn,

Jump to, or create a Workspace.

Call ‘get_name’ to obtain a Workspace name and check to see if there is currently a Workspace with that name being managed by the WindowManager. If there is no existing workspace with the given name, create it with the supplied available layouts. If a matching Workspace does already exist then simply switch focus to it. This action is most useful when combined with the DefaultWorkspace hook that allows for auto populating named Workspaces when first focusing them.

If you just want to dynamically select an existing workspace then you can use switch_to_workspace to select from known workspace names.