penrose::extensions::actions

Function toggle_fullscreen

pub fn toggle_fullscreen<X: XConn>() -> Box<dyn KeyEventHandler<X>>

Toggle the fullscreen state of the currently focused window.

NOTE: You will need to make use of add_ewmh_hooks for this action to work correctly. Additionally you can use the ClearFsPropOnStartup hook to prevent existing clients from having an inconsistent fullscreen state after restarting penrose while they are fullscreen.