penrose::extensions

Module hooks

Hook implementations and helpers for adding to your Penrose window manager

Re-exports

pub use ewmh::add_ewmh_hooks;

pub use named_scratchpads::add_named_scratchpads;

pub use named_scratchpads::NamedScratchPad;

pub use named_scratchpads::ToggleNamedScratchPad;

pub use startup::SpawnOnStartup;

pub use window_swallowing::WindowSwallowing;

Modules

default_workspaces

Configure workspaces to auto-spawn a set of windows if they are empty when they gain focus

ewmh

EWMH compliance for Penrose

manage

Manage hooks for common manage actions

named_scratchpads

Support for managing multiple floating scratchpad programs that can be toggled on or off on the active workspace.

startup

Startup hooks for direct adding to your penrose config.

window_swallowing

Window swallowing in the style of XMonad.Hooks.WindowSwallowing