Module penrose::extensions::hooks::manage

Manage hooks for common manage actions

Manage hooks should not trigger a refresh directly: that is handled by penrose itself when the manage hook is called.

Structs

DefaultTiled

Perform no additional actions when managing a new client.

FloatingCentered

Float clients in the center of the screen.

FloatingFixed

Float clients at a fixed position on the screen.

FloatingRelative

Float clients at a relative position to the current screen.

SetWorkspace

Move the specified client to the named workspace.