penrose::extensions::hooks

Module startup

Startup hooks for direct adding to your penrose config.

Structs

ClearFsPropOnStartup

Remove _NET_WM_STATE_FULLSCREEN property from existing clients, not present in client_set, on window manager startup.

SpawnOnStartup

Spawn a client program on window manager startup