Struct penrose::extensions::util::dmenu::DMenu

pub struct DMenu { /* private fields */ }

A wrapper around the suckless dmenu program for creating dynamic menus in penrose.

Implementations

impl DMenu

pub fn new(config: &DMenuConfig, screen_index: usize) -> Self

Create a new DMenu command which can be triggered and reused by calling the run method for a basic dmenu prompt, or the build_menu for more advanced selection menus.

pub fn run(&self) -> Result<()>

Used for launching regular old DMenu with no menu matching via the dmenu_run wrapper script.

pub fn build_menu( &self, param_choices: Vec<impl Into<String>> ) -> Result<MenuMatch>

Run this DMenu command and return the selected choice.

Example

let screen_index = 0;
let dmenu = DMenu::new(&DMenuConfig::default(), screen_index);

let choices = vec!["some", "choices", "to", "pick", "from"];

match dmenu.build_menu(choices).unwrap() {
    MenuMatch::Line(i, s) => println!("matched '{}' on line '{}'", s, i),
    MenuMatch::UserInput(s) => println!("user input: '{}'", s),
    MenuMatch::NoMatch => println!("no match"),
}