penrose

Module pure

Side effect free management of internal window manager state

Modules

geometry

Geometry primitives

Structs

Diff

The internal diff state from the last time that we refreshed pure State.

Screen

A wrapper around a single Workspace that includes the physical screen size as a Rect.

ScreenClients

Used in laying out visible_client_positions

Snapshot

A summary of the information required to update the X server state from our own internal pure State

Stack

A Stack can be thought of as a linked list with a hole punched in it to mark a single element that currently holds focus (though in practice it is implemented using a VecDeque for efficiency purposes). By convention, the main element is the first element in the stack (regardless of focus). Focusing operations do not reorder the elements of the stack or the resulting Vec that can be obtained from calling Stack::flatten.

StackSet

The side-effect free internal state representation of the window manager.

Workspace

A wrapper around a Stack of windows belonging to a single “workspace” or virtual desktop. When this workspace is active on a given screen, the windows contained in its stack will be positioned using the active layout of its LayoutStack.

Enums

Position

A position within a Stack.

RelativePosition

A relative position along the horizontal and vertical axes