Struct penrose::pure::geometry::Point

pub struct Point {
    pub x: u32,
    pub y: u32,
}

An x,y coordinate pair

Fields
x: u32

An absolute x coordinate relative to the root window

y: u32

An absolute y coordinate relative to the root window

Implementations

impl Point

pub fn new(x: u32, y: u32) -> Self

Create a new Point.