Struct penrose::pure::geometry::Rect

pub struct Rect {
    pub x: u32,
    pub y: u32,
    pub w: u32,
    pub h: u32,
}

An X window / screen position: top left corner + extent

Fields
x: u32

The x-coordinate of the top left corner of this rect

y: u32

The y-coordinate of the top left corner of this rect

w: u32

The width of this rect

h: u32

The height of this rect

Implementations

impl Rect

pub const fn new(x: u32, y: u32, w: u32, h: u32) -> Rect

Create a new Rect.

pub fn corners(&self) -> (Point, Point, Point, Point)

The four corners of this Rect in Point form returned in clockwise order from the top left corner.

let r = Rect::new(0, 0, 100, 200);
let corners = r.corners();

assert_eq!(
    corners,
    (
        Point { x: 0, y: 0 },
        Point { x: 100, y: 0 },
        Point { x: 100, y: 200 },
        Point { x: 0, y: 200 },
    )
);

pub fn midpoint(&self) -> Point

The midpoint of this rectangle.

Odd side lengths will lead to a truncated point towards the top left corner in order to maintain integer coordinates.

let r = Rect::new(0, 0, 100, 200);

assert_eq!(r.midpoint(), Point { x: 50, y: 100 });

pub fn shrink_in(&self, border: u32) -> Self

Shrink width and height by the given pixel border, maintaining the current x and y coordinates. The resulting Rect will always have a minimum width and height of 1.

let r = Rect::new(0, 0, 100, 200);

assert_eq!(r.shrink_in(10), Rect::new(0, 0, 80, 180));
assert_eq!(r.shrink_in(50), Rect::new(0, 0, 1, 100));
assert_eq!(r.shrink_in(100), Rect::new(0, 0, 1, 1));

pub fn scale_w(&self, factor: f64) -> Self

Create a new Rect with width equal to factor x self.w

let r = Rect::new(0, 0, 30, 40);

assert_eq!(r.scale_w(1.5), Rect::new(0, 0, 45, 40));
assert_eq!(r.scale_w(0.5), Rect::new(0, 0, 15, 40));

pub fn scale_h(&self, factor: f64) -> Self

Create a new Rect with height equal to factor x self.h

let r = Rect::new(0, 0, 30, 40);

assert_eq!(r.scale_h(1.5), Rect::new(0, 0, 30, 60));
assert_eq!(r.scale_h(0.5), Rect::new(0, 0, 30, 20));

pub fn resize(&mut self, dw: i32, dh: i32)

Update the width and height of this Rect by specified deltas.

Minimum size is clamped at 1x1.

Panics

This function will panic if one of the supplied deltas overflows i32::MAX.

let mut r = Rect::new(0, 0, 100, 200);

r.resize(20, 30);
assert_eq!(r, Rect::new(0, 0, 120, 230));

r.resize(-40, -50);
assert_eq!(r, Rect::new(0, 0, 80, 180));

pub fn reposition(&mut self, dx: i32, dy: i32)

Update the position of this Rect by specified deltas.

Minimum (x, y) coordinates are clamped at (0, 0)

Panics

This function will panic if one of the supplied deltas overflows i32::MAX.

let mut r = Rect::new(0, 0, 100, 200);

r.reposition(20, 30);
assert_eq!(r, Rect::new(20, 30, 100, 200));

r.reposition(-40, -20);
assert_eq!(r, Rect::new(0, 10, 100, 200));

pub fn contains(&self, other: &Rect) -> bool

Check whether this Rect contains other as a sub-Rect

pub fn is_larger_than(&self, other: &Rect) -> bool

Check whether this Rect is physically larger than other regardless of position.

pub fn contains_point<P>(&self, p: P) -> bool
where P: Into<Point>,

Check whether this Rect contains p

pub fn centered_in(&self, enclosing: &Rect) -> Option<Self>

Center this Rect inside of enclosing.

Returns None if this Rect can not fit inside enclosing

pub fn as_rows(&self, n_rows: u32) -> Vec<Rect>

Split this Rect into evenly sized rows.

pub fn as_columns(&self, n_columns: u32) -> Vec<Rect>

Split this Rect into evenly sized columns.

pub fn split_at_width(&self, new_width: u32) -> Option<(Self, Self)>

Divides this rect into two columns where the first has the given width.

Returns None if new_width is out of bounds

pub fn split_at_height(&self, new_height: u32) -> Option<(Self, Self)>

Divides this rect into two rows where the first has the given height.

Returns None if new_height is out of bounds

pub fn split_at_width_perc(&self, perc: f32) -> Option<(Self, Self)>

Divide this rect into two columns where the first takes up perc% of the current width.

Returns None if perc is not between 0.0 and 1.0

pub fn split_at_height_perc(&self, perc: f32) -> Option<(Self, Self)>

Divide this rect into two rows where the first takes up perc% of the current height.

Returns None if perc is not between 0.0 and 1.0

pub fn split_at_mid_width(&self) -> (Self, Self)

Divides this rect into two columns along its midpoint.

pub fn split_at_mid_height(&self) -> (Self, Self)

Divides this rect into two rows along its midpoint.