penrose
Module util

Utility functions for use in other parts of penrose

Functions

notify

Use notify-send to display a message to the user

print_layout_result

Run a given Layout for a stack of n clients and print a simple ASCII rendering of the resulting client positions.

spawn

Run an external command

spawn_for_output

Run an external command and return its output.

spawn_for_output_with_args

Run an external command with arguments and return its output.

spawn_with_args

Run an external command with the specified command line arguments