penrose::util

Function spawn_for_output

pub fn spawn_for_output<S: Into<String>>(cmd: S) -> Result<String>

Run an external command and return its output.

std::process::Command::output will not work within penrose due to the way that signal handling is set up. Use this function if you need to access the output of a process that you spawn.