penrose

Module x

Logic for interacting with the X server

Re-exports

pub use atom::Atom;

pub use event::XEvent;

pub use property::Prop;

pub use property::WindowAttributes;

Modules

atom

Data types for working with X atoms

event

Data types for working with X events

property

Data types for working with X window properties

query

Queries against client windows

Structs

XConnState

Transient state needed for all XConn impls in order to track expected map/unmap events coming from the Xserver.

Enums

ClientAttr

Attributes for an X11 client window (not all are curently implemented)

ClientConfig

On screen configuration options for X clients (not all are curently implemented)

WinType

A window type to be specified when creating a new window in the X server

Traits

XConn

A handle on a running X11 connection that we can use for issuing X requests.