Module penrose::x::event

Data types for working with X events

Structs

ClientMessage

A client message that needs to be parsed and handled based on its type

ConfigureEvent

A configure request or notification when a client changes position or size

ExposeEvent

A notification that a window has become visible

PointerChange

A notification that the mouse pointer has entered or left a window

PropertyEvent

A property change on a known client

ResizeRequestEvent

A client is being resized

Enums

ClientEventMask

Event masks used when sending client events

ClientMessageData

The raw data contained in a ClientMessage

ClientMessageKind

Known common client message formats.

XEvent

Wrapper around the low level X event types that correspond to request / response data when communicating with the X server itself.