Enum penrose::x::event::ClientMessageData

pub enum ClientMessageData {
    U8([u8; 20]),
    U16([u16; 10]),
    U32([u32; 5]),
}

The raw data contained in a ClientMessage

Variants

U8([u8; 20])

Slice of u8

U16([u16; 10])

Slice of u16

U32([u32; 5])

Slice of u32

Implementations

impl ClientMessageData

pub fn as_usize(&self) -> Vec<usize>

Convert this client message into a single data format

The number of raw values will be maintained but this allows you to have a consistant interface without needing to match on the variant or cast all the time.

impl ClientMessageData

pub fn as_u8(&self) -> Vec<u8>

Convert this client message into a single data format

The number of raw values will be maintained but this allows you to have a consistant interface without needing to match on the variant or cast all the time.

impl ClientMessageData

pub fn as_u16(&self) -> Vec<u16>

Convert this client message into a single data format

The number of raw values will be maintained but this allows you to have a consistant interface without needing to match on the variant or cast all the time.

impl ClientMessageData

pub fn as_u32(&self) -> Vec<u32>

Convert this client message into a single data format

The number of raw values will be maintained but this allows you to have a consistant interface without needing to match on the variant or cast all the time.