Module penrose::x::property

Data types for working with X window properties

Structs

WindowAttributes

Window Attributes honoured by penose.

WmHints

Client requested hints about information other than window geometry.

WmHintsFlags

Possible flags that can be set in a WmHints client property

WmNormalHints

Client requested hints about window geometry.

WmNormalHintsFlags

Possible flags that can be set in a WmNormalHints client property

Enums

MapState

The mapping states a window can be in

Prop

Know property types that should be returnable by XConn impls when they check window properties.

WindowClass

The input class for a window

WmState

Possible valid values for setting the WM_STATE property on a client.