Struct penrose::x::property::WmNormalHints

pub struct WmNormalHints { /* private fields */ }

Client requested hints about window geometry.

See the ICCCM spec for further details or the Xlib manual for more details of the data fromat but note that Penrose does not honour the following hints:

gravity

increment

aspect ratio

Implementations

impl WmNormalHints

pub fn new(
flags: WmNormalHintsFlags,
base: Option<Rect>,
min: Option<Rect>,
max: Option<Rect>,
user_specified: Option<Rect>
) -> Self

Create a new instance from component parts

pub fn apply_to(&self, r: Rect) -> Rect

Apply these size hints to a given Rect.

NOTE: currently only the max size is respected

pub fn try_from_bytes(raw: &[u32]) -> Result<Self>

Try to construct a WmNormalHints instance from raw bytes.

This method expects a slice of 18 u32s corresponding to the C struct layout shown below.

typedef struct {
    long flags;                /* marks which fields in this structure are defined */
    int x, y;                  /* Obsolete */
    int width, height;         /* Obsolete */
    int min_width, min_height;
    int max_width, max_height;
    int width_inc, height_inc;
    struct {
           int x;              /* numerator */
           int y;              /* denominator */
    } min_aspect, max_aspect;
    int base_width, base_height;
    int win_gravity;
    /* this structure may be extended in the future */
} XSizeHints;