penrose::x::query

Trait Query

pub trait Query<X: XConn> {
    // Required method
    fn run(&self, id: Xid, x: &X) -> Result<bool>;

    // Provided methods
    fn and<Other>(self, other: Other) -> AndQuery<X>
       where Self: Sized + 'static,
             Other: Query<X> + 'static { ... }
    fn or<Other>(self, other: Other) -> OrQuery<X>
       where Self: Sized + 'static,
             Other: Query<X> + 'static { ... }
    fn not(self) -> NotQuery<X>
       where Self: Sized + 'static { ... }
}

A query to be run against client windows for identifying specific windows or programs.

Required Methods

fn run(&self, id: Xid, x: &X) -> Result<bool>

Run this query for a given window ID.

Provided Methods

fn and<Other>(self, other: Other) -> AndQuery<X>
where Self: Sized + 'static, Other: Query<X> + 'static,

Combine this query with another query using a logical AND.

This follows typical short-circuiting behavior, i.e. if the first query returns false, the second query will not be run.

fn or<Other>(self, other: Other) -> OrQuery<X>
where Self: Sized + 'static, Other: Query<X> + 'static,

Combine this query with another query using a logical OR.

This follows typical short-circuiting behavior, i.e. if the first query returns true, the second query will not be run.

fn not(self) -> NotQuery<X>
where Self: Sized + 'static,

Apply a logical NOT to this query.

This will invert the result of the query.