Trait penrose::x::XConn

pub trait XConn {

    fn root(&self) -> Xid;
    fn screen_details(&self) -> Result<Vec<Rect>>;
    fn cursor_position(&self) -> Result<Point>;
    fn grab(
        &self,
        key_codes: &[KeyCode],
        mouse_states: &[MouseState]
    ) -> Result<()>;
    fn next_event(&self) -> Result<XEvent>;
    fn flush(&self);
    fn intern_atom(&self, atom: &str) -> Result<Xid>;
    fn atom_name(&self, xid: Xid) -> Result<String>;
    fn client_geometry(&self, client: Xid) -> Result<Rect>;
    fn existing_clients(&self) -> Result<Vec<Xid>>;
    fn map(&self, client: Xid) -> Result<()>;
    fn unmap(&self, client: Xid) -> Result<()>;
    fn kill(&self, client: Xid) -> Result<()>;
    fn focus(&self, client: Xid) -> Result<()>;
    fn get_prop(&self, client: Xid, prop_name: &str) -> Result<Option<Prop>>;
    fn list_props(&self, client: Xid) -> Result<Vec<String>>;
    fn get_wm_state(&self, client: Xid) -> Result<Option<WmState>>;
    fn get_window_attributes(&self, client: Xid) -> Result<WindowAttributes>;
    fn set_wm_state(&self, client: Xid, wm_state: WmState) -> Result<()>;
    fn set_prop(&self, client: Xid, name: &str, val: Prop) -> Result<()>;
    fn delete_prop(&self, client: Xid, prop_name: &str) -> Result<()>;
    fn set_client_attributes(
        &self,
        client: Xid,
        attrs: &[ClientAttr]
    ) -> Result<()>;
    fn set_client_config(&self, client: Xid, data: &[ClientConfig]) -> Result<()>;
    fn send_client_message(&self, msg: ClientMessage) -> Result<()>;
    fn warp_pointer(&self, id: Xid, x: i16, y: i16) -> Result<()>;

}

A handle on a running X11 connection that we can use for issuing X requests.

XConn is intended as an abstraction layer to allow for communication with the underlying display system (assumed to be X) using whatever mechanism the implementer wishes. In theory, it should be possible to write an implementation that allows penrose to run on systems not using X as the windowing system but X idioms and high level event types / client interations are assumed.

Required Methods

fn root(&self) -> Xid

The ID of the window manager root window.

fn screen_details(&self) -> Result<Vec<Rect>>

Ask the X server for the dimensions of each currently available screen.

fn cursor_position(&self) -> Result<Point>

Ask the X server for the current (x, y) coordinate of the mouse cursor.

fn grab(&self, key_codes: &[KeyCode], mouse_states: &[MouseState]) -> Result<()>

Grab the specified key and mouse states, intercepting them for processing within the window manager itself.

fn next_event(&self) -> Result<XEvent>

Block and wait for the next event from the X server so it can be processed.

fn flush(&self)

Flush any pending events to the X server.

fn intern_atom(&self, atom: &str) -> Result<Xid>

Look up the Xid of a given Atom name. If it is not currently interned, intern it.

fn atom_name(&self, xid: Xid) -> Result<String>

Look up the string name of a given Atom by its Xid.

fn client_geometry(&self, client: Xid) -> Result<Rect>

Look up the current dimensions and position of a given client window.

fn existing_clients(&self) -> Result<Vec<Xid>>

Ask the X server for the IDs of all currently known client windows

fn map(&self, client: Xid) -> Result<()>

Map the given client window to the screen with its current geometry, making it visible.

fn unmap(&self, client: Xid) -> Result<()>

Unmap the given client window from the screen, hiding it.

fn kill(&self, client: Xid) -> Result<()>

Kill the given client window, closing it.

fn focus(&self, client: Xid) -> Result<()>

Set X input focus to be held by the given client window.

fn get_prop(&self, client: Xid, prop_name: &str) -> Result<Option<Prop>>

Look up a specific property on a given client window.

fn list_props(&self, client: Xid) -> Result<Vec<String>>

List the known property names set for a given client.

fn get_wm_state(&self, client: Xid) -> Result<Option<WmState>>

Get the current WmState for a given client window.

fn get_window_attributes(&self, client: Xid) -> Result<WindowAttributes>

Request the WindowAttributes for a given client window from the X server.

fn set_wm_state(&self, client: Xid, wm_state: WmState) -> Result<()>

Set the current WmState for a given client window.

fn set_prop(&self, client: Xid, name: &str, val: Prop) -> Result<()>

Set a specific property on a given client window.

fn delete_prop(&self, client: Xid, prop_name: &str) -> Result<()>

Delete a property for a given client window.

fn set_client_attributes(&self, client: Xid, attrs: &[ClientAttr]) -> Result<()>

Set one or more ClientAttr for a given client window.

fn set_client_config(&self, client: Xid, data: &[ClientConfig]) -> Result<()>

Set the ClientConfig for a given client window.

fn send_client_message(&self, msg: ClientMessage) -> Result<()>

Send a ClientMessage to a given client.

fn warp_pointer(&self, id: Xid, x: i16, y: i16) -> Result<()>

Reposition the mouse cursor to the given (x, y) coordinates within the specified window. This method should not be called directly: use warp_pointer_to_window or warp_pointer_to_screen instead.