Type Alias penrose::x11rb::RustConn

pub type RustConn = Conn<RustConnection>;

A pure rust based connection to the X server using a [RustConnection].

Aliased Type

struct RustConn { /* private fields */ }

Implementations

impl Conn<RustConnection>

pub fn new() -> Result<Self>

Construct an X11rbConnection backed by the x11rb backend using [x11rb::rust_connection::RustConnection].

impl<C> Conn<C>where C: Connection,

pub fn connection(&self) -> &C

Get a handle to the underlying connection.

pub fn create_window(&self, ty: WinType, r: Rect, managed: bool) -> Result<Xid>

Create and map a new window to the screen with the specified WinType.

pub fn destroy_window(&self, id: Xid) -> Result<()>

Destroy the window identified by the given Xid.