Module penrose::xcb

Helpers and utilities for using XCB as a back end for penrose

This is the reference implementation of the core Penrose traits that are required to back the WindowManager when talking to the X server. The code in this module is build on top of the xcb, pango and cairo C libraries and strives to be as simple as possible (as opposed to optimising for performance).

Available features

xcb_draw: adds pango and cairo as build dependencies and provides implementations of the Draw and DrawContext traits.

xcb_keysyms: if enabled, this will include the XKeySym enum which is based on the contents of X11/keysymdef.h and allows for parsing X keycode events as their corresponding utf-8 representation or action.

C level documentation

Docs for the underlying xcb C library can be found here.

Modules

api
A wrapper around the underlying xcb api layer that only exposes Penrose types

draw
API layer implementing Draw and DrawContext using XCB, pango and cairo.

helpers
XCB based helper functions

keysyms
Auto generated Keysym enum for use with xcb keycodes

xconn
API wrapper for talking to the X server using XCB

Structs

Api
A connection to the X server using the XCB C API

XcbConnection
Handles communication with an X server via the XCB library.

XcbDraw
An XCB based Draw implementation backed by pango and cairo

XcbDrawContext
An XCB based drawing context using pango and cairo

Enums

XKeySym
X keysym mappings: auto generated from X11/keysymdef.h

XcbError
Enum to store the various ways that operations can fail inside of the XCB implementations of penrose traits.

Traits

XcbApi
An abstraction layer for talking to the X server using the XCB api.

Functions

new_xcb_backed_window_manager
Construct a penrose WindowManager backed by the default xcb backend.

Type Definitions

Result
Result type for fallible methods using XCB

XcbGenericEvent
A generic event type returned by the xcb library

XcbHooks
Helper type for when you are defining your Hook vector in your main.rs when using the default XCB impls