Struct penrose::xcb::XcbConnection

pub struct XcbConnection { /* fields omitted */ }

Handles communication with an X server via the XCB library.

XcbConnection is a minimal implementation that does not make use of the full asyc capabilities of the underlying C XCB library.

Implementations

impl XcbConnection

pub fn new() -> Result<Self>

Establish a new connection to the running X server. Fails if unable to connect

pub fn xcb_connectction(&self) -> &Connection

Get a handle on the underlying [XCB Connection][::xcb::Connection] used by Api to communicate with the X server.

pub fn api(&self) -> &Api

Get a handle on the underlying Api to communicate with the X server.

pub fn api_mut(&mut self) -> &mut Api

Get a mutable handle on the underlying Api to communicate with the X server.

pub fn known_atoms(&self) -> &HashMap<Atom, u32>

The current interned Atom values known to the underlying Api connection