Trait penrose::xcb::XcbApi

pub trait XcbApi {
    fn hydrate(&mut self) -> Result<()>;

    fn atom(&self, name: &str) -> Result<u32>;

    fn known_atom(&self, atom: Atom) -> u32;

    fn delete_prop(&self, id: WinId, prop: Atom);

    fn get_atom_prop(&self, id: WinId, atom: Atom) -> Result<u32>;

    fn get_str_prop(&self, id: WinId, name: &str) -> Result<String>;

    fn replace_prop(&self, id: WinId, prop: Atom, val: PropVal<'_>);

    fn create_window(
        &self, 
        ty: WinType, 
        r: Region, 
        managed: bool
    ) -> Result<WinId>;

    fn configure_window(&self, id: WinId, conf: &[WinConfig]);

    fn current_clients(&self) -> Result<Vec<WinId>>;

    fn destroy_window(&self, id: WinId);

    fn focused_client(&self) -> Result<WinId>;

    fn map_window(&self, id: WinId);

    fn mark_focused_window(&self, id: WinId);

    fn send_client_event(&self, id: WinId, atom_name: &str) -> Result<()>;

    fn set_window_attributes(&self, id: WinId, attrs: &[WinAttr]);

    fn unmap_window(&self, id: WinId);

    fn window_geometry(&self, id: WinId) -> Result<Region>;

    fn current_screens(&self) -> Result<Vec<Screen>>;

    fn screen_sizes(&self) -> Result<Vec<Region>>;

    fn cursor_position(&self) -> Point;

    fn grab_keys(&self, keys: &[&KeyCode]);

    fn grab_mouse_buttons(&self, states: &[&MouseState]);

    fn ungrab_keys(&self);

    fn ungrab_mouse_buttons(&self);

    fn flush(&self) -> bool;

    fn root(&self) -> WinId;

    fn set_randr_notify_mask(&self) -> Result<()>;

    fn wait_for_event(&self) -> Result<XEvent>;

    fn poll_for_event(&self) -> Result<Option<XEvent>>;

    fn warp_cursor(&self, id: WinId, x: usize, y: usize);
}

An abstraction layer for talking to the X server using the XCB api.

This has been written to be a reasonably close mapping to the underlying C API, but provides several quality of life changes that make consuming the API nicer to work with in Penrose code.

Required methods

fn hydrate(&mut self) -> Result<()>

Hydrate this XcbApi to restore internal state following serde deserialization

fn atom(&self, name: &str) -> Result<u32>

Intern an atom by name, returning the corresponding id.

Can fail if the atom name is not a known X atom or if there is an issue with communicating with the X server. For known atoms that are included in the Atom enum, the XcbApi::known_atom method should be used instead.

fn known_atom(&self, atom: Atom) -> u32

Fetch the id value of a known Atom variant.

This operation is expected to always succeed as known atoms should either be interned on init of the implementing struct or statically assigned a value in the implementation.

fn delete_prop(&self, id: WinId, prop: Atom)

Delete a known property from a window

fn get_atom_prop(&self, id: WinId, atom: Atom) -> Result<u32>

Fetch an Atom property for a given window

fn get_str_prop(&self, id: WinId, name: &str) -> Result<String>

Fetch an String property for a given window

fn replace_prop(&self, id: WinId, prop: Atom, val: PropVal<'_>)

Replace a property value on a window.

See the documentation for the C level XCB API for the correct property type for each prop.

fn create_window(&self, ty: WinType, r: Region, managed: bool) -> Result<WinId>

Create a new client window

fn configure_window(&self, id: WinId, conf: &[WinConfig])

Apply a set of config options to a window

fn current_clients(&self) -> Result<Vec<WinId>>

The list of currently active clients known to the X server

fn destroy_window(&self, id: WinId)

Destroy the X server state for a given window

fn focused_client(&self) -> Result<WinId>

The client that the X server currently considers to be focused

fn map_window(&self, id: WinId)

Send a XEvent::MapRequest for the target window

fn mark_focused_window(&self, id: WinId)

Mark the given window as currently having focus in the X server state

fn send_client_event(&self, id: WinId, atom_name: &str) -> Result<()>

Send an event to a client

fn set_window_attributes(&self, id: WinId, attrs: &[WinAttr])

Set attributes on the target window

fn unmap_window(&self, id: WinId)

Unmap the target window

fn window_geometry(&self, id: WinId) -> Result<Region>

Find the current size and position of the target window

fn current_screens(&self) -> Result<Vec<Screen>>

Query the randr API for current outputs and return the details as penrose Screen structs.

fn screen_sizes(&self) -> Result<Vec<Region>>

Query the randr API for current outputs and return the size of each screen

fn cursor_position(&self) -> Point

The current (x, y) position of the cursor relative to the root window

fn grab_keys(&self, keys: &[&KeyCode])

Register intercepts for each given KeyCode

fn grab_mouse_buttons(&self, states: &[&MouseState])

Register intercepts for each given MouseState

fn ungrab_keys(&self)

Drop all active intercepts for key combinations

fn ungrab_mouse_buttons(&self)

Drop all active intercepts for mouse states

fn flush(&self) -> bool

Flush pending actions to the X event loop

fn root(&self) -> WinId

The current root window ID

fn set_randr_notify_mask(&self) -> Result<()>

Set a pre-defined notify mask for randr events to subscribe to

fn wait_for_event(&self) -> Result<XEvent>

Block until the next event from the X event loop is ready then return it.

This method should handle all of the mapping of xcb events to penrose XEvent instances, returning an Error when the event channel from the X server is closed.

fn poll_for_event(&self) -> Result<Option<XEvent>>

Return the next event from the X event loop if there is one.

This method should handle all of the mapping of xcb events to penrose XEvent instances, returning None if there is no pending event and an error if the connection to the X server is closed.

fn warp_cursor(&self, id: WinId, x: usize, y: usize)

Move the cursor to the given (x, y) position inside the specified window.

Loading content...