Module penrose::xcb::xconn

API wrapper for talking to the X server using XCB

The crate used by penrose for talking to the X server is rust-xcb, which is a set of bindings for the C level XCB library that are autogenerated from an XML spec. The XML files can be found here and are useful as reference for how the API works. Sections have been converted and added to the documentation of the method calls and enums present in this module.

EWMH Xlib manual

Structs

XcbConnection
Handles communication with an X server via the XCB library.