Crate penrose

Penrose: a library for building your very own tiling window manager

Modules

builtin

Built-in functionality for penrose.

core

Core data structures and user facing functionality for the window manager

extensions

Extensions to the base behaviour of Penrose

macros

Utility macros

pure

Side effect free management of internal window manager state

util

Utility functions for use in other parts of penrose

x

Logic for interacting with the X server

x11rb

Helpers and utilities for using x11rb as a back end for penrose

Macros

custom_error

Quickly create a crate::Error::Custom

map

Make creating a pre-defined HashMap a little less verbose

simple_transformer

Quickly define a LayoutTransformer from a single element tuple struct and a transformation function: fn(Rect, Vec<(Xid, Rect)>) -> Vec<(Xid, Rect)>.

stack

Create a Stack containing the arguments. The only required element is the focus, it is not possible to create an empty Stack;

Structs

Color

A simple RGBA based color

Xid

An X11 ID for a given resource

Enums

Error

Type Definitions

Result