Module penrose::builtin::actions::floating

Actions for manipulating floating windows.

Structs

MouseDragHandler

A simple mouse event handler for dragging a window

MouseResizeHandler

A simple mouse event handler for resizing a window

Functions

float_all

Float all windows in their current tiled position

float_focused

Move the currently focused windo to the floating layer in its current on screen position

reposition

Move a currently floating window by a given (x, y) delta

resize

Resize a currently floating window by a given (width, height) delta

sink_all

Sink all floating windows back into their tiled positions

sink_focused

Sink the current window back into tiling mode if it was floating