penrose::builtin

Module layout

Built-in layouts.

Modules

messages

Built-in layout messages.

transformers

Built-in layout transformers.

Structs

CenteredMain

A simple Layout with a main and secondary side regions.

Grid

A simple grid layout that places windows in the smallest nxn grid that will contain all window present on the workspace.

MainAndStack

A simple Layout with main and secondary regions.

Monocle

A simple monolce layout that gives the maximum available space to the currently focused client and unmaps all other windows.