penrose

Module core

Core data structures and user facing functionality for the window manager

Re-exports

pub use conn::WinId;

Modules

bindings

Setting up and responding to user defined key/mouse bindings

conn

A platform agnostic backing connection

hooks

Traits for writing and composing hooks.

layout

Layouts for positioning client windows on the screen within a given workspace.

Structs

Config

The user specified config options for how the window manager should run

State

Mutable internal state for the window manager

WindowManager

A top level struct holding all of the state required to run as an X11 window manager.

Type Aliases

ClientSet

The pure client state information for the window manager

ClientSpace

The pure client state information for a single Workspace