Module penrose::core::bindings

Setting up and responding to user defined key/mouse bindings

Structs

KeyCode
A key press and held modifiers

ModifierKeyIter

MouseEvent
A mouse movement or button event

MouseState
A mouse state specification indicating the button and modifiers held

Enums

KeyPress
Abstraction layer for working with key presses

ModifierKey
Known modifier keys for bindings

MouseButton
Known mouse buttons for binding actions

MouseEventKind
The types of mouse events represented by a MouseEvent

Type Definitions

KeyBindings
User defined key bindings

KeyCodeMask
A u16 X key-code bitmask

KeyCodeValue
A u8 X key-code enum value

KeyEventHandler
Some action to be run by a user key binding

MouseBindings
User defined mouse bindings

MouseEventHandler
An action to be run in response to a mouse event