Trait penrose::core::bindings::MouseEventHandler

pub trait MouseEventHandler<X>where
    X: XConn,{
    // Required method
    fn call(
        &mut self,
        evt: &MouseEvent,
        state: &mut State<X>,
        x: &X
    ) -> Result<()>;
}

An action to be run in response to a mouse event

Required Methods

fn call(&mut self, evt: &MouseEvent, state: &mut State<X>, x: &X) -> Result<()>

Call this handler with the current window manager state and mouse state