Module penrose::core::hooks

Traits for writing and composing hooks

Structs

ComposedEventHook

The result of composing two event hooks using then

ComposedManageHook

The result of composing two manage hooks using then

ComposedStateHook

The result of composing two state hooks using then

Traits

EventHook

Handle an XEvent, return true if default event handling should be run afterwards.

ManageHook

Action to run when a new client becomes managed.

StateHook

An arbitrary action that can be run and modify State