Trait penrose::core::hooks::EventHook

pub trait EventHook<X> where
    X: XConn, {
    fn call(&mut self, event: &XEvent, state: &mut State<X>, x: &X) -> Result<bool>;

    fn boxed(self) -> Box<dyn EventHook<X>>
    where
        Self: Sized + 'static,
    { ... }
    fn then<H>(self, next: H) -> ComposedEventHook<X>
    where
        H: EventHook<X> + 'static,
        Self: Sized + 'static,
    { ... }
    fn then_boxed(self, next: Box<dyn EventHook<X>>) -> Box<dyn EventHook<X>>
    where
        Self: Sized + 'static,
        X: 'static,
    { ... }
}

Handle an XEvent, return true if default event handling should be run afterwards.

This hook is called before incoming XEvents are processed by the default event handling logic.

Required Methods

fn call(&mut self, event: &XEvent, state: &mut State<X>, x: &X) -> Result<bool>

Run this hook

Provided Methods

fn boxed(self) -> Box<dyn EventHook<X>> where
Self: Sized + 'static,

Convert to a trait object

fn then<H>(self, next: H) -> ComposedEventHook<X> where
H: EventHook<X> + 'static,
Self: Sized + 'static,

Compose this hook with another EventHook. The second hook will be skipped if this one returns false.

fn then_boxed(self, next: Box<dyn EventHook<X>>) -> Box<dyn EventHook<X>> where
Self: Sized + 'static,
X: 'static,

Compose this hook with a boxed EventHook. The second hook will be skipped if this one returns false.