Trait penrose::core::hooks::ManageHook

pub trait ManageHook<X> where
    X: XConn, {
    fn call(&mut self, client: Xid, state: &mut State<X>, x: &X) -> Result<()>;

    fn boxed(self) -> Box<dyn ManageHook<X>>
    where
        Self: Sized + 'static,
    { ... }
    fn then<H>(self, next: H) -> ComposedManageHook<X>
    where
        H: ManageHook<X> + 'static,
        Self: Sized + 'static,
    { ... }
    fn then_boxed(self, next: Box<dyn ManageHook<X>>) -> Box<dyn ManageHook<X>>
    where
        Self: Sized + 'static,
        X: 'static,
    { ... }
}

Action to run when a new client becomes managed.

Manage hooks should not trigger refreshes of state directly: they are called immediately before a refresh is run by main window manager logic.

Required Methods

fn call(&mut self, client: Xid, state: &mut State<X>, x: &X) -> Result<()>

Run this hook

Provided Methods

fn boxed(self) -> Box<dyn ManageHook<X>> where
Self: Sized + 'static,

Convert to a trait object

fn then<H>(self, next: H) -> ComposedManageHook<X> where
H: ManageHook<X> + 'static,
Self: Sized + 'static,

Compose this hook with another ManageHook.

fn then_boxed(self, next: Box<dyn ManageHook<X>>) -> Box<dyn ManageHook<X>> where
Self: Sized + 'static,
X: 'static,

Compose this hook with a boxed ManageHook.