Module penrose::core::layout

Layouts for positioning client windows on the screen within a given workspace.

Re-exports

pub use messages::IntoMessage;

pub use messages::Message;

pub use transformers::LayoutTransformer;

Modules

messages

transformers

Traits

Layout

A Layout is responsible for positioning a Stack of clients in a given coordinate space denoting the dimensions of users display.

Type Definitions

LayoutStack

A stack of Layout options for use on a particular [Workspace].