Module penrose::core::layout

Layouts for positioning client windows on the screen within a given workspace.

Structs

Message

A dynamically typed message to be sent to a Layout for processing.

Traits

IntoMessage

Marker trait for a type that can be sent as a Message.

Layout

A Layout is responsible for positioning a Stack of clients in a given coordinate space denoting the dimensions of users display.

LayoutTransformer

A wrapper round another Layout that is able to intercept and modify both the positions being returned by the inner layout and messages being sent to it.

Type Aliases

LayoutStack

A stack of Layout options for use on a particular crate::pure::Workspace.