penrose::core::layout

Trait IntoMessage

pub trait IntoMessage: Any {
    // Provided method
    fn into_message(self) -> Message
       where Self: Sized { ... }
}

Marker trait for a type that can be sent as a Message.

The impl_message macro can be used to easily implement this trait and mark a type as being usable as a layout message:

use penrose::impl_message;

struct MyMessage;
impl_message!(MyMessage);

Provided Methods

fn into_message(self) -> Message
where Self: Sized,

Wrap this value as a dynamically typed message for sending to a layout