Trait penrose::core::layout::IntoMessage

pub trait IntoMessage: Any {
    fn into_message(self) -> Message
    where
        Self: Sized,
    { ... }
}

Marker trait for a type that can be sent as a crate::core::layout::Message

Provided Methods

fn into_message(self) -> Messagewhere
Self: Sized,

Wrap this value as a dynamically typed message for sending to a layout