Trait penrose::core::layout::LayoutTransformer

pub trait LayoutTransformer: Clone + Sized + 'static {
    fn transformed_name(&self) -> String;
    fn inner_mut(&mut self) -> &mut Box<dyn Layout>;

    fn swap_inner(&mut self, new: Box<dyn Layout>) -> Box<dyn Layout> { ... }
    fn unwrap(&mut self) -> Box<dyn Layout> { ... }
    fn transform_initial(&self, r: Rect) -> Rect { ... }
    fn transform_positions(
        &mut self,
        _r: Rect,
        positions: Vec<(Xid, Rect)>
    ) -> Vec<(Xid, Rect)> { ... }
    fn run_transform<F>(
        &mut self,
        f: F,
        r: Rect
    ) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>)
    where
        F: FnOnce(Rect, &mut Box<dyn Layout>) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>),
    { ... }
    fn passthrough_message(&mut self, m: &Message) -> Option<Box<dyn Layout>> { ... }
}

A wrapper round another Layout that is able to intercept and modify both the positions being returned by the inner layout and messages being sent to it.

Required Methods

fn transformed_name(&self) -> String

The same as Layout::name but for LayoutTransformer itself.

fn inner_mut(&mut self) -> &mut Box<dyn Layout>

Provide a mutable reference to the Layout wrapped by this transformer.

Provided Methods

fn swap_inner(&mut self, new: Box<dyn Layout>) -> Box<dyn Layout>

Replace the currently wrapped Layout with a new one.

fn unwrap(&mut self) -> Box<dyn Layout>

Remove the inner Layout from this LayoutTransformer.

The default implementation of this method simply swaps the built-in Monocle layout for the inner layout using LayoutTransformer::swap_inner. If a better implementation is possible then it should be implemented but this default will give the correct behaviour when LayoutTransformer::passthrough_message receives an UnwrapTransformer message.

fn transform_initial(&self, r: Rect) -> Rect

Modify the initial Rect that will be passed to the inner Layout.

The default implementation of this method leaves the initial Rect unchanged.

fn transform_positions(
&mut self,
_r: Rect,
positions: Vec<(Xid, Rect)>
) -> Vec<(Xid, Rect)>

Optionally modify any of the positions returned by the inner Layout before they are applied by the window manager. The dimensions of the screen being layed out are avaiable as r.

The default implementation of this method leaves the positions returned by the inner layout unchanged.

fn run_transform<F>(
&mut self,
f: F,
r: Rect
) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>)where
F: FnOnce(Rect, &mut Box<dyn Layout>) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>),

Apply the LayoutTransformer to its wrapped inner Layout.

fn passthrough_message(&mut self, m: &Message) -> Option<Box<dyn Layout>>

Pass a message on to the wrapped inner Layout.

The default implementation of this method will return Some(inner_layout) if it receives an UnwrapTransformer Message using the unwrap method of this trait.