Type Definition penrose::core::layout::LayoutStack

pub type LayoutStack = Stack<Box<dyn Layout>>;

A stack of Layout options for use on a particular crate::pure::Workspace.

The Stack itself acts as a Layout, deferring all operations to the currently focused Layout.

Implementations

impl LayoutStack

pub fn run_and_replace<F>(&mut self, f: F) -> Vec<(Xid, Rect)> where
F: FnOnce(&mut Box<dyn Layout>) -> (Option<Box<dyn Layout>>, Vec<(Xid, Rect)>),

Run the currently focused Layout and return the positions it generates.

If the layout being run wants to be replaced with a new layout, swap it out for the new one in its current position in the Stack.

pub fn handle_message<M>(&mut self, m: M) where
M: IntoMessage,

Send the given Message to the currently active Layout.

pub fn broadcast_message<M>(&mut self, m: M) where
M: IntoMessage,

Send the given Message to every Layout in this stack rather that just the currently active one.