Module penrose::extensions::actions

Helpers and pre-defined actions for use in user defined key bindings

Enums

FullScreenAction

The possible valid actions to use when manipulating full screen state

Functions

create_or_switch_to_workspace

Jump to, or create, a crate::pure::Workspace

dmenu_focus_client

Use DMenu to dynamically select and focus a client window.

dmenu_focus_tag

Use DMenu to dynamically select and focus a client window.

focus_or_spawn

Focus a client with the given class as WM_CLASS or spawn the program with the given command if no such client exists.

set_fullscreen_state

Set the fullscreen state of a particular client

toggle_fullscreen

Toggle the fullscreen state of the currently focused window.