Module penrose::extensions::hooks::ewmh

EWMH compliance for Penrose

The primary use of this extension is to provide support for external status bars / panels and fullscreen windows.

See details of the spec here: https://specifications.freedesktop.org/wm-spec/wm-spec-latest.html

Constants

EWMH_SUPPORTED_ATOMS

The set of Atoms this extension adds support for.

WM_NAME

The WM_NAME that will be set for the X server

Functions

add_ewmh_hooks

Add the required hooks to manage EWMH compliance to an existing crate::core::Config.

event_hook

Intercept messages from external applications and handle them.

refresh_hook

Notify external clients of the current status of workspaces and clients

startup_hook

Advertise EWMH support to the X server