Module penrose::extensions::hooks::named_scratchpads

Support for managing multiple floating scratchpad programs that can be toggled on or off on the active workspace.

Structs

NamedScratchPad

A toggle-able client program that can be shown and hidden via a keybinding.

ToggleNamedScratchPad

Toggle the visibility of a NamedScratchPad.

Constants

NSP_TAG

The tag used for a placeholder Workspace that holds scratchpad windows when they are currently hidden.

Functions

add_named_scratchpads

Add the required hooks to manage EWMH compliance to an existing crate::core::Config.

event_hook

Remove destroyed clients from internal scratchpad state

manage_hook

Store clients matching NamedScratchPad queries and run the associated ManageHook.