penrose::extensions

Module util

Utility extensions for use in the penrose window manager

Modules

debug

Debugging utilities for diagnosing issues with penrose.

dmenu

A simple wrapper for suckless’ dmenu tool for providing quick text based menus.

Enums

NotifyLevel

A notification level when calling notify-send

Functions

notify_send

Send a notification using the notify-send command line program

notify_send_custom

Send a notification using the notify-send command line program

update_monitors_via_xrandr

Detect the current monitor set up and arrange the monitors if needed using xrandr.