Macro penrose::simple_transformer

macro_rules! simple_transformer {
    (
        $(#[$struct_docs:meta])*
        $t:ident,
        $f:ident,
        $prefix:expr
    ) => { ... };
}

Quickly define a LayoutTransformer from a single element tuple struct and a transformation function: fn(Rect, Vec<(Xid, Rect)>) -> Vec<(Xid, Rect)>.

The struct must have a single field which is a Box<dyn Layout>.

Example

fn my_transformation_function(r: Rect, positions: Vec<(Xid, Rect)>) -> Vec<(Xid, Rect)> {
    // transformation implementation goes here
    positions
}

simple_transformer!(MyTransformer, my_transformation_function, "MyTransform");