Struct penrose::pure::geometry::Rect

pub struct Rect {
    pub x: u32,
    pub y: u32,
    pub w: u32,
    pub h: u32,
}

An X window / screen position: top left corner + extent

Fields
x: u32

The x-coordinate of the top left corner of this rect

y: u32

The y-coordinate of the top left corner of this rect

w: u32

The width of this rect

h: u32

The height of this rect

Implementations

impl Rect

pub fn new(x: u32, y: u32, w: u32, h: u32) -> Rect

Create a new Rect.

pub fn corners(&self) -> (Point, Point, Point, Point)

The four corners of this Rect in Point form returned in clockwise order from the top left corener.

pub fn midpoint(&self) -> Point

The midpoint of this rectangle

pub fn shrink_in(&self, border: u32) -> Self

Shrink width and height by the given pixel border, maintaining the current x and y coordinates.

pub fn scale_w(&self, factor: f64) -> Self

Create a new Rect with width equal to factor x self.w

pub fn scale_h(&self, factor: f64) -> Self

Create a new Rect with height equal to factor x self.h

pub fn resize(&mut self, dw: i32, dh: i32)

Update the width and height of this Rect by specified deltas.

Minimum size is clamped at 0x0

pub fn reposition(&mut self, dx: i32, dy: i32)

Update the position of this Rect by specified deltas.

Minimum (x, y) coordinates are clamped at (0, 0)

pub fn contains(&self, other: &Rect) -> bool

Check whether this Rect contains other as a sub-Rect

pub fn is_larger_than(&self, other: &Rect) -> bool

Check whether this Rect is physically larger than other regardless of position.

pub fn contains_point<P>(&self, p: P) -> bool where
P: Into<Point>,

Check whether this Rect contains p

pub fn centered_in(&self, enclosing: &Rect) -> Option<Self>

Center this Rect inside of enclosing.

Returns None if this Rect can not fit inside enclosing

pub fn as_rows(&self, n_rows: u32) -> Vec<Rect>

Split this Rect into evenly sized rows.

pub fn as_columns(&self, n_columns: u32) -> Vec<Rect>

Split this Rect into evenly sized columns.

pub fn split_at_width(&self, new_width: u32) -> Option<(Self, Self)>

Divides this rect into two columns where the first has the given width.

Returns None if new_width is out of bounds

pub fn split_at_height(&self, new_height: u32) -> Option<(Self, Self)>

Divides this rect into two rows where the first has the given height.

Returns None if new_height is out of bounds