penrose::util

Function print_layout_result

pub fn print_layout_result<L: Layout>(
    l: &mut L,
    n_clients: u32,
    screen_width: u32,
    screen_height: u32,
)

Run a given Layout for a stack of n clients and print a simple ASCII rendering of the resulting client positions.

This is provided as a helper function to make it easier to develop your own layout functions and see the results without needing to make use of them directly in your window manager. The output shown below is obtained by running the MainAndStack layout with a screen size of (40, 15) and 4 clients.

A note on screen dimensions

In order to print out the ASCII representation of your layout, the “screen” dimensions being used will be significantly smaller than a typical real-life monitor. You may observe some unexpected behaviour from your layout when running this function if you have any hard coded values relating to screen size or individual client sizes.

.........................................
.                       .               .
.                       .               .
.                       .               .
.                       .               .
.                       .................
.                       .               .
.                       .               .
.                       .               .
.                       .               .
.                       .................
.                       .               .
.                       .               .
.                       .               .
.                       .               .
.........................................