Trait penrose::x::XConnExt

pub trait XConnExt: XConn + Sized {

    fn kill_focused(&self, state: &mut State<Self>) -> Result<()> { ... }
    fn manage(&self, id: Xid, state: &mut State<Self>) -> Result<()> { ... }
    fn unmanage(&self, client: Xid, state: &mut State<Self>) -> Result<()> { ... }
    fn reveal(
        &self,
        client: Xid,
        cs: &ClientSet,
        mapped: &mut HashSet<Xid>
    ) -> Result<()> { ... }
    fn hide(
        &self,
        client: Xid,
        mapped: &mut HashSet<Xid>,
        pending_unmap: &mut HashMap<Xid, usize>
    ) -> Result<()> { ... }
    fn modify_and_refresh<F>(&self, state: &mut State<Self>, f: F) -> Result<()>
    where
        F: FnMut(&mut ClientSet),
    { ... }
    fn refresh(&self, state: &mut State<Self>) -> Result<()> { ... }
    fn client_should_float(
        &self,
        client: Xid,
        floating_classes: &[String]
    ) -> Result<bool> { ... }
    fn set_client_border_color<C>(&self, id: Xid, color: C) -> Result<()>
    where
        C: Into<Color>,
    { ... }
    fn set_initial_properties(
        &self,
        client: Xid,
        config: &Config<Self>
    ) -> Result<()> { ... }
    fn position_client(&self, client: Xid, r: Rect) -> Result<()> { ... }
    fn position_clients(
        &self,
        border: u32,
        positions: &[(Xid, Rect)]
    ) -> Result<()> { ... }
    fn restack<'a, I>(&self, ids: I) -> Result<()>
    where
        I: Iterator<Item = &'a Xid>,
    { ... }
    fn set_active_client(
        &self,
        client: Xid,
        state: &mut State<Self>
    ) -> Result<()> { ... }
    fn warp_pointer_to_window(&self, id: Xid) -> Result<()> { ... }
    fn warp_pointer_to_screen(
        &self,
        state: &mut State<Self>,
        screen_index: usize
    ) -> Result<()> { ... }
    fn window_title(&self, id: Xid) -> Result<String> { ... }
    fn client_supports_protocol(&self, id: Xid, proto: &str) -> Result<bool> { ... }
    fn window_pid(&self, id: Xid) -> Option<u32> { ... }
    fn query(&self, query: &dyn Query<Self>, id: Xid) -> Result<bool> { ... }
    fn query_or(&self, default: bool, query: &dyn Query<Self>, id: Xid) -> bool { ... }

}

Extended functionality for XConn impls in order to run the window manager.

Provided Methods

fn kill_focused(&self, state: &mut State<Self>) -> Result<()>

Kill the focused client if there is one

fn manage(&self, id: Xid, state: &mut State<Self>) -> Result<()>

Establish the window manager state for the given client window and refresh the current X state.

fn unmanage(&self, client: Xid, state: &mut State<Self>) -> Result<()>

Remove the window manager state for the given client window and refresh the current X state.

fn reveal(
&self,
client: Xid,
cs: &ClientSet,
mapped: &mut HashSet<Xid>
) -> Result<()>

Display a client on the screen by mapping it and setting its WmState to Normal This is idempotent if the client is already visible.

fn hide(
&self,
client: Xid,
mapped: &mut HashSet<Xid>,
pending_unmap: &mut HashMap<Xid, usize>
) -> Result<()>

Hide a client by unmapping it and setting its WmState to Iconic

fn modify_and_refresh<F>(&self, state: &mut State<Self>, f: F) -> Result<()> where
F: FnMut(&mut ClientSet),

Apply a pure function that modifies a ClientSet and then handle refreshing the WindowManager state and associated X11 calls.

This is the main logic that drives what the user will see on the screen in terms of window placement, focus and borders. Everything is driven from a diff of the pure ClientSet state before and after some mutating operation that was carried out by f.

fn refresh(&self, state: &mut State<Self>) -> Result<()>

Refresh the current X server state based on a diff of the current state against the state when we last refreshed.

fn client_should_float(
&self,
client: Xid,
floating_classes: &[String]
) -> Result<bool>

Check whether or not the given client should be assigned floating status or not.

fn set_client_border_color<C>(&self, id: Xid, color: C) -> Result<()> where
C: Into<Color>,

Update the border color of the given client window.

fn set_initial_properties(&self, client: Xid, config: &Config<Self>) -> Result<()>

Set the initial window properties for a newly managed window.

fn position_client(&self, client: Xid, r: Rect) -> Result<()>

Update the geometry of a given client based on the given Rect.

fn position_clients(&self, border: u32, positions: &[(Xid, Rect)]) -> Result<()>

Restack and set the geometry for an ordered list of client windows and their associated positions. The provided positions are shrunk by the current border size in order to position the windows correctly within the frame given by the border.

See restack for details of stacking order is determined.

fn restack<'a, I>(&self, ids: I) -> Result<()> where
I: Iterator<Item = &'a Xid>,

Restack the given windows in, each one above the last.

fn set_active_client(&self, client: Xid, state: &mut State<Self>) -> Result<()>

Update the currently focused client and refresh the X state.

fn warp_pointer_to_window(&self, id: Xid) -> Result<()>

Warp the mouse cursor to the center of the given client window.

fn warp_pointer_to_screen(
&self,
state: &mut State<Self>,
screen_index: usize
) -> Result<()>

Warp the mouse cursor to the center of the given screen.

fn window_title(&self, id: Xid) -> Result<String>

Request the title of a given client window following ICCCM/EWMH standards.

fn client_supports_protocol(&self, id: Xid, proto: &str) -> Result<bool>

Check to see if a given client window supports a particular protocol or not

fn window_pid(&self, id: Xid) -> Option<u32>

Request a window’s PID via the _NET_WM_PID property.

NOTE: Not all programs set this property.

fn query(&self, query: &dyn Query<Self>, id: Xid) -> Result<bool>

Run the provided Query, returning the result.

fn query_or(&self, default: bool, query: &dyn Query<Self>, id: Xid) -> bool

Run the provided Query, returning the result or a default value if there were any errors encountered when communicating with the X server.