Module penrose::x11rb

Helpers and utilities for using x11rb as a back end for penrose

Docs for the X11 core protocol can be found here. x11rb is a thin facade over this. For X11 extensions, there are usually separate documentations. For example, the RandR extension is documented in 2.

This module contains the code for talking to the X11 server using the x11rb crate, which offers an implementation of the X11 protocol in safe Rust. The actual protocol bindings are autogenerated from an XML spec. The XML files can be found in its xcb-proto-{something} subfolder and are useful as a reference for how the API works. x11rb also offers some explanation on how the XML is turned into Rust code.

The original implementation of this was by @psychon (Uli Schlachter). Re-write for the new 0.3.0 API by @sminez (Innes Anderson-Morrison).

Modules

conversions

Conversions to Penrose types from X11rb types

Structs

Conn

Handles communication with an X server via the x11rb crate.

Type Definitions

RustConn

A pure rust based connection to the X server using a [RustConnection].

XcbConn

An C based connection to the X server using an [XCBConnection].