Struct penrose::x11rb::Conn

pub struct Conn<C: Connection> { /* private fields */ }

Handles communication with an X server via the x11rb crate.

Implementations

impl Conn<RustConnection>

pub fn new() -> Result<Self>

Construct an X11rbConnection backed by the x11rb backend using [x11rb::rust_connection::RustConnection].

impl Conn<XCBConnection>

pub fn new() -> Result<Self>

Construct an X11rbConnection backed by the x11rb backend using [x11rb::xcb_ffi::XCBConnection].

impl<C> Conn<C> where
C: Connection,

pub fn connection(&self) -> &C

Get a handle to the underlying connection.

pub fn create_window(&self, ty: WinType, r: Rect, managed: bool) -> Result<Xid>

Create and map a new window to the screen with the specified WinType.